Fitted values from a stackedsdm object

Fitted values from a stackedsdm object

Usage

# S3 method for stackedsdm
fitted(object, ...)

Arguments

object

An object of class stackedsdm

...

Not used

Value

A matrix of fitted values.

Details

Extracts the fitted values from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

library(mvabund)
data(spider)
X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2)
fitted(fit0)
#>             Alopacce    Alopcune    Alopfabr     Arctlute     Arctperi
#> units1  18.239439075  5.72944913  1.84824965 1.733787e-01 2.614114e-01
#> units2   0.574937738 10.91758845  0.11883678 1.233708e+00 1.000000e-06
#> units3  12.796390651  8.37422532  1.22135400 4.663690e-01 4.803516e-02
#> units4   6.369090246  7.23571889  0.67265832 1.084879e+00 1.948166e-03
#> units5   1.847246864 10.66403979  0.41045498 1.489015e+00 1.764620e-05
#> units6   0.053339718  7.54439444  0.01714693 2.512629e+00 1.000000e-06
#> units7   1.790316488 12.78328159  0.25495713 6.165215e+00 1.728380e-05
#> units8   0.060496800  3.34807121  0.07847928 7.634772e-03 1.000000e-06
#> units9   4.804652178  2.38867237  0.78547053 3.629855e-01 3.940469e-02
#> units10  5.796189060  1.68662396  1.97337649 2.807266e-02 1.632348e-01
#> units11 10.674743518  2.50073273  1.61594333 1.023714e-01 2.057460e-01
#> units12 12.354251711  3.47387850  1.48946423 1.710713e-01 1.554745e-01
#> units13  1.719075301 50.81514469  0.13419156 7.425636e+00 1.733456e-04
#> units14  0.753297790  5.58325940  0.12583476 1.083397e+01 1.000000e-06
#> units15  0.023626731  7.29239238  0.02332829 1.276512e-01 1.000000e-06
#> units16  0.002244303  0.88788163  0.02450816 3.254817e-03 1.000000e-06
#> units17  0.031399335  9.82017726  0.02509812 1.558768e-01 1.000000e-06
#> units18  0.034405328  7.53034101  0.04516223 3.022311e-02 1.000000e-06
#> units19  0.005982467  1.21620548  0.02518152 9.466707e-03 1.000000e-06
#> units20  0.002112249  0.99585409  0.02105357 5.098833e-03 1.000000e-06
#> units21  0.005684769  1.61539778  0.02747983 7.436178e-03 1.000000e-06
#> units22  2.693446952  0.09773166 11.70871816 2.485391e-05 4.473446e+00
#> units23 24.195787923  0.50063982 15.31524154 1.779486e-04 8.855710e+00
#> units24 13.068803874  0.82511545  5.99586115 1.865888e-03 2.476450e+00
#> units25  3.834123624  1.20844732  3.32399352 2.604614e-03 3.131571e-04
#> units26  8.297089369  0.20577260 44.17303581 2.375144e-06 1.785605e+01
#> units27 22.962508395  0.33002954 20.61323125 5.384577e-05 2.287569e+00
#> units28 22.580750360  0.33332523 21.15718899 4.945857e-05 2.175068e+00
#>             Auloalbi    Pardlugu    Pardmont     Pardnigr     Pardpull
#> units1  6.246029e+00  0.14673395 46.21546509 4.571731e+00 2.286804e+01
#> units2  4.839317e+01  2.88728364  9.62564065 3.397070e+01 7.796638e+01
#> units3  1.040089e+01  0.17670987 40.47461115 1.104517e+01 4.515780e+01
#> units4  1.299844e+01  0.32927856 41.79161499 2.031875e+01 7.041935e+01
#> units5  8.287397e+00  0.93614429  9.07087888 9.866695e+01 4.405246e+01
#> units6  2.879000e+00  3.32587128  7.04213262 9.804684e+00 1.237658e+01
#> units7  9.714575e+00  0.56385795 18.66309570 1.234824e+02 1.007958e+02
#> units8  4.502981e+00 26.25187948  0.78179195 1.340972e+00 7.890039e-01
#> units9  6.398908e-02  0.06793954 49.95754379 2.081384e+00 1.986247e+00
#> units10 1.972793e-02  0.09422697 15.72709170 9.316283e-01 2.757519e-01
#> units11 3.903659e-01  0.10576271 47.14012274 1.600086e+00 3.460346e+00
#> units12 1.339434e+00  0.12270998 53.86315317 2.670330e+00 9.188397e+00
#> units13 2.015290e+01  0.17807669 16.04219884 4.142190e+01 1.234677e+02
#> units14 1.464170e+01  1.81532567 30.15026027 1.389557e+02 1.634871e+02
#> units15 8.244225e-01  8.61005347  0.85836505 3.892906e+00 9.310314e-01
#> units16 6.175724e-05  8.36622883  0.07483028 3.066005e-01 3.975826e-04
#> units17 7.693405e-01  4.93906712  0.96911828 3.524352e+00 1.011545e+00
#> units18 6.270717e-01  8.65363817  0.45495797 2.639564e+00 3.944188e-01
#> units19 3.439665e-03 11.89787290  0.29309531 5.059937e-01 9.006105e-03
#> units20 6.073417e-05  7.29219102  0.08257814 3.682140e-01 4.854220e-04
#> units21 6.111447e-03 16.82592849  0.17765208 7.813532e-01 1.033129e-02
#> units22 1.294083e-06  0.06920151  1.46477591 9.372339e-03 2.979863e-05
#> units23 5.580253e-02  0.29363456  9.45280688 8.838901e-02 5.688250e-02
#> units24 2.219924e-02  0.13388839 13.45905784 2.083026e-01 9.380578e-02
#> units25 1.484514e+00  3.61200153  3.28647894 3.463452e+00 9.000806e-01
#> units26 6.238258e-05  0.27056389  0.42166194 1.940265e-02 9.175248e-05
#> units27 1.277356e-01  0.83371593  6.61331698 7.766916e-02 5.116620e-02
#> units28 1.240644e-01  0.86658930  6.06029510 7.947422e-02 4.773396e-02
#>            Trocterr    Zoraspin
#> units1   18.3352321  1.74400341
#> units2   90.3951831 20.75715483
#> units3   32.3616070  3.56653846
#> units4   43.8164845  7.68985486
#> units5  138.7180969 28.55463853
#> units6   57.6057630 15.77690371
#> units7  148.8111236 36.27503647
#> units8   22.9857922  3.11312349
#> units9   11.2139988  1.43401223
#> units10   7.9268024  0.63335099
#> units11   9.3564798  0.97926407
#> units12  12.6292245  1.39331046
#> units13 130.8191121 11.06876236
#> units14 126.6461033 73.16976425
#> units15  51.0227314  7.19512558
#> units16  14.6242186  1.58643682
#> units17  50.8350811  5.63734903
#> units18  44.8805177  4.20044127
#> units19  15.4748074  2.36951595
#> units20  16.3574371  1.83033895
#> units21  21.9123589  3.03423657
#> units22   0.6045805  0.02023397
#> units23   1.7354742  0.08741095
#> units24   3.0348509  0.18053844
#> units25  16.5802466  2.36683229
#> units26   1.0923949  0.02147195
#> units27   1.5198742  0.09073516
#> units28   1.5598137  0.09142142

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
fitted(fit0)
#>          Alopacce   Alopcune  Alopfabr  Arctlute    Arctperi  Auloalbi
#> units1  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units2  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units3  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units4  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units5  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units6  0.8635980 0.43106837 0.8265953 0.5916564  0.88208574 1.7510773
#> units7  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units8  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units9  0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units10 0.9602934 0.13499534 0.9801575 0.3729749  9.05551052 0.7160559
#> units11 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units12 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units13 0.7885040 0.58584687 0.6542511 0.7099879  0.35145427 2.4932161
#> units14 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units15 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units16 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units17 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units18 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units19 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units20 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units21 0.4379158 0.89957172 0.1097251 1.2171196  0.02314281 7.0862156
#> units22 0.9306739 0.23802342 0.9464921 0.4567645  3.25610041 1.0605149
#> units23 0.9176213 0.28018845 0.9273941 0.4870765  2.35427463 1.2011487
#> units24 0.9306739 0.23802342 0.9464921 0.4567645  3.25610041 1.0605149
#> units25 0.9469193 0.18261553 0.9667202 0.4141730  5.33648900 0.8772650
#> units26 0.9756768 0.07914657 0.9916957 0.3133469 21.81483980 0.5108911
#> units27 0.8635980 0.43106837 0.8265953 0.5916564  0.88208574 1.7510773
#> units28 0.9399572 0.20673274 0.9585553 0.4332241  4.25285269 0.9571539
#>          Pardlugu  Pardmont  Pardnigr   Pardpull  Trocterr Zoraspin
#> units1  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units2  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units3  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units4  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units5  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units6  0.7451292 10.255083  5.644297  4.5056558 13.969898 4.636483
#> units7  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units8  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units9  6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units10 0.1821669  6.638133  2.344651  1.1576482  6.063784 3.191637
#> units11 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units12 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units13 1.3000532 12.178051  7.986693  7.7083295 19.427203 5.373657
#> units14 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units15 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units16 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units17 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units18 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units19 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units20 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units21 6.7387742 20.240920 22.286760 37.7025694 51.499104 8.312125
#> units22 0.3381893  8.035432  3.448655  2.1027545  8.748512 3.760469
#> units23 0.4114807  8.537155  3.897437  2.5408149  9.826669 3.961187
#> units24 0.3381893  8.035432  3.448655  2.1027545  8.748512 3.760469
#> units25 0.2508310  7.327193  2.862279  1.5761156  7.328996 3.474068
#> units26 0.1070307  5.632903  1.682817  0.6930472  4.424932 2.771960
#> units27 0.7451292 10.255083  5.644297  4.5056558 13.969898 4.636483
#> units28 0.2877438  7.644479  3.118154  1.7993245  7.950058 3.602837