Calculate residuals from a stackedsdm object

Calculate residuals from a stackedsdm object

Usage

# S3 method for stackedsdm
residuals(object, type = "dunnsmyth", seed = NULL, ...)

Arguments

object

An object of class stackedsdm;

type

Determined what type of residuals to calculate. The current options include Dunn-Smyth residuals (default; "dunnsmyth"), raw response residuals ("response") or probability integral transform residuals ("PIT");

seed

For Dunn-Smyth and PIT residuals applied to discrete responses, random jittering is added, and the seed can be used to seed to jittering.

...

not used

Value

A matrix of residuals

Details

Calculated the residuals from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2) 
residuals(fit0)
#>           Alopacce   Alopcune    Alopfabr    Arctlute     Arctperi    Auloalbi
#> units1   0.8869130  0.9097481 -0.81190640 -2.45240675 -0.835215570 -0.52369416
#> units2  -0.3434584 -0.9285419 -1.26693704 -0.42205320  0.741866980 -0.80464653
#> units3   0.4313962  1.3074959  0.86678335  1.79784186 -0.030260400 -0.07527700
#> units4  -1.3737627  0.2032565 -0.94815803  0.58265579  0.136670486  1.53104295
#> units5  -0.2249609  1.0077469 -0.75287691  0.69160086 -1.325299285  0.32981717
#> units6   1.2342610  0.1397448 -0.51832000  1.19782843  0.210745882  1.35685155
#> units7   0.4541911 -0.1540515 -0.07875371  1.02757341  0.878821230  1.10964521
#> units8   1.1194705  1.7054761 -1.07451275 -0.51540063 -1.123259882  1.04308675
#> units9  -1.4974539 -0.3217752 -1.08696726 -1.40761266 -0.095125030 -1.01135075
#> units10 -0.9124340 -0.9291554 -0.13651135 -0.91132321  0.452484170 -2.32148272
#> units11  0.8141393 -0.3619516  0.55302503  0.01265863 -0.546221068  0.71011149
#> units12  0.7485332  1.7797721 -1.49261880 -0.35877439  0.053307725 -1.18958079
#> units13  0.7846018  0.1793624  1.18981912 -0.65183651 -0.834690121 -0.03163515
#> units14 -0.5820088 -0.3127485 -0.07760650 -1.37415144 -2.203783467 -1.72132593
#> units15  0.2353525 -1.3801016 -0.07344529  0.64601729 -1.777065865 -0.77896618
#> units16  0.7457426  1.2860228 -1.42969103  0.03938784 -2.146817581 -1.01807766
#> units17 -2.3138827 -1.3192417  0.13066577 -1.35996525 -0.341142443 -2.78163936
#> units18  1.7110392 -1.1234257  1.66199877  0.57863999  2.377504214 -0.38660130
#> units19 -0.6435101  0.1551491 -0.59305613  0.03664649  0.281658741  1.39314165
#> units20 -0.8077387  1.0961270  0.26967302  0.43097525  0.744021557  1.43312121
#> units21  0.1799414  0.0880132 -2.48250717 -0.73268108 -0.356184550 -0.27696720
#> units22  2.0169289 -0.8664772  0.66036782  1.25171761  0.029916194  0.54477225
#> units23 -0.6512914  0.3671523  0.25917139 -1.10523471 -0.442893838  0.02689504
#> units24 -0.2527131 -1.4458136  1.91403315 -1.10024146  1.346873448 -0.09592040
#> units25  1.7866250  0.4226503  1.51610162 -1.26487559  0.729725975  0.38967748
#> units26 -1.2796156 -2.2173195 -1.36672857 -0.53997974 -0.005356787  0.12131200
#> units27  0.7067297  0.2317339 -0.31029183 -0.56705692  1.122302264 -1.05889610
#> units28 -0.7188934  0.3740507 -0.12715910 -0.51187489 -0.699737127 -1.09361580
#>            Pardlugu    Pardmont     Pardnigr    Pardpull    Trocterr   Zoraspin
#> units1  -0.03658536  0.61257576  1.466450081  1.56716658  2.43332162  1.1752814
#> units2  -0.96016358 -1.60663332 -0.437807151 -1.06106616 -0.33479244 -0.5697288
#> units3   1.09821768 -0.13165379  0.832924005  0.15068667  1.52117279 -0.5382376
#> units4   0.74555610 -1.49509043  0.708895213  0.73736814  1.44838844  1.9508947
#> units5   0.68174644 -1.12964618  0.660100230  1.29224723 -0.47125970 -0.3061021
#> units6  -1.77293861  0.79679802  1.537359299  1.47370512  0.35348658  1.3255543
#> units7   0.34490744  0.89550013 -0.001874459  0.24860959 -0.18120421 -0.5852381
#> units8   1.73333383  0.89465532  0.756313011  0.04429322  0.66440951  0.2299663
#> units9  -2.51972622 -0.48558475  0.052232665 -0.13262248 -1.59998751 -1.3148528
#> units10 -0.08524834  0.69380499 -0.247828594  0.24560589 -1.71246678 -1.5524661
#> units11 -0.38495318  1.26790406 -0.470603829 -0.71382043 -0.74749084 -0.3758293
#> units12  0.49713902  1.07448909 -0.358277984 -0.09382804  0.24689604 -0.3493802
#> units13  1.46214405  0.79896110  0.590744385 -0.76421536 -0.27900583  1.1921296
#> units14  0.77110688 -0.62209717 -1.379489837 -1.21519550 -0.28563580 -0.6194211
#> units15 -0.49186296 -1.79652094 -1.631311475 -0.52361755 -0.86271987 -0.6177669
#> units16 -0.25759508  0.01903475  2.117588839  1.13942959  1.35330678  1.4412978
#> units17 -1.08513324 -1.31010470 -1.021248832 -0.90614352 -0.96798062 -0.5538538
#> units18 -0.68962440  0.40561843 -1.862376158  0.26443400 -0.65198383 -1.4515741
#> units19  0.15145770  0.59199972  0.613649406 -1.34802685  0.77706557  0.6294775
#> units20  1.15852574 -0.39965326 -0.114574480  0.09305082  0.67404046  0.3649610
#> units21  0.11557620  1.21494144  0.097609395  2.54792863 -0.11211713 -0.1483822
#> units22 -0.96438037  0.37612208 -0.708069044 -0.27089710  1.08916286 -1.2388342
#> units23  2.32179501 -0.20369615  1.050249268 -0.82774796 -0.04437463  0.3455151
#> units24 -1.02177187 -1.23419878  0.331172979 -0.54901038 -1.84374756 -0.4161752
#> units25 -1.18018352  1.90532972  0.953462683 -1.62676695  0.10960545  1.4336785
#> units26 -0.15357508 -1.33155093  1.888072967  0.37074764 -0.12415368  0.4824713
#> units27 -1.47071921 -1.08525508 -0.765487226 -0.11159739 -2.16366490 -0.6351997
#> units28 -1.48763895  0.24783708  0.943306420  1.40154223  0.30990543 -0.2085068

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
residuals(fit0)
#>            Alopacce    Alopcune     Alopfabr      Arctlute   Arctperi
#> units1   0.33315166 -0.45704288 -1.889192180 -1.600195e+00  0.7689570
#> units2  -0.03238627 -0.73521463  0.499899271 -4.532242e-01 -0.5579963
#> units3   1.95983764 -0.37241773  1.543043324  1.003620e+00  0.7771763
#> units4   1.94638160 -0.04463378 -0.834266810  8.764060e-01  0.6568944
#> units5   1.30079752  2.11396086 -0.574175312  9.273521e-01 -1.3985308
#> units6  -1.22305811  1.69484926 -2.535723657  1.996624e+00  0.1106641
#> units7   0.78865978 -0.72376643 -0.036310641  2.029272e+00  0.7765786
#> units8  -0.51004066 -0.17320466  0.489190424  5.371593e-01  1.4750978
#> units9   1.18696886 -0.01258353 -0.647164359 -1.631216e+00  0.2916488
#> units10  0.46435962  0.09670373  0.864056016 -7.016631e-01 -1.0708067
#> units11  0.18909723  1.21529769  2.433430470 -1.346101e+00 -0.8260949
#> units12  0.56444408 -1.09589536 -1.459589659 -1.056693e+00 -0.5004468
#> units13  1.23242167  0.06079179  1.982137455  1.153123e+00 -0.7469696
#> units14 -0.29648187 -0.40458829 -0.703529372  7.770033e-01  0.1889098
#> units15  0.12083750 -2.24705994 -0.016884204 -4.006578e-01 -0.5646070
#> units16 -0.46343847 -0.23871488 -1.510713784 -7.482922e-01 -0.9340590
#> units17 -0.73792026 -1.48068108 -0.287714178  4.505429e-05 -0.8477073
#> units18 -1.63086716 -0.66597608  0.577354893 -5.758727e-01 -0.7377657
#> units19 -0.62883677 -0.38725909 -1.244216517  7.281634e-02  1.3022767
#> units20  0.03832647  1.38451676  0.079720827 -2.557669e-01 -1.8966587
#> units21 -0.42258167  1.65747371  0.464667179  7.339848e-02  0.6045169
#> units22 -0.14723836 -0.19833534 -1.329784104  5.010819e-01  0.5140904
#> units23  0.55666097  0.41207716  0.785301172 -4.906617e-01  1.2728020
#> units24  0.49650431 -1.05018555  0.051647675  6.751196e-01  0.7924462
#> units25  0.84897452  1.00801365 -1.015652293 -4.270188e-01 -0.8805808
#> units26 -0.66168757  0.93049851 -0.007206241  1.083674e-01  0.3401880
#> units27 -0.25239891 -0.39828360 -0.502267110 -1.846864e+00  1.6173977
#> units28 -0.69584637 -0.60500416  0.193376103  7.748523e-02 -0.3254672
#>           Auloalbi    Pardlugu     Pardmont     Pardnigr    Pardpull
#> units1   0.4694496 -0.61287678  1.303526038  0.507360931  0.83777683
#> units2   1.5099443 -0.28943512 -0.709355200  0.585577005  0.76664032
#> units3   0.8223772 -0.26154660  0.758253810  0.653858466  0.84439053
#> units4   1.3570441 -0.54418692  0.034369545  0.842501844  1.17760074
#> units5   0.8196218 -0.31876740 -0.390795253  1.754190786  1.07660596
#> units6   1.3587255 -0.06902292  0.581050758  1.572623206  1.62025380
#> units7   1.0838564 -0.39003101  0.787625040  1.451758351  1.23666667
#> units8   0.6874896  2.52840173 -0.472015559  0.029059229 -0.12395577
#> units9  -0.8423904 -2.31661204  0.693243957 -0.130964231 -0.15592110
#> units10 -0.5560140 -0.46865679  1.409504766 -0.249727403  0.47138874
#> units11  0.1566779 -2.54800246  1.736610217 -0.370801197 -0.15895067
#> units12 -0.7810788 -0.93943008  1.747861917 -0.101151614  0.28368781
#> units13  1.9237552  0.11851652  0.992086142  1.818349640  2.07252015
#> units14  0.4723337 -0.02878498  0.362204707  0.589507568  1.04851471
#> units15 -0.2385922  0.39816470 -1.416580881 -1.199517866 -0.74537763
#> units16 -0.5265815  0.42644771 -1.836024171 -0.005167325 -0.32998110
#> units17 -1.4930840 -0.10957762 -1.095801514 -0.900085293 -0.27675640
#> units18 -0.6514745  0.25993646 -1.377161945 -0.389507627 -0.63671475
#> units19 -0.4844318  0.88026005 -1.503426299 -0.192296097 -0.71460896
#> units20 -0.6022671  0.98023302 -2.121659124 -0.375909144 -1.48275711
#> units21 -0.5647897  1.11425704 -0.586480208 -1.632015732 -0.14160417
#> units22 -0.6157938 -1.26219655 -0.009180321 -1.767364310 -0.61955667
#> units23 -1.6800203  1.57695614  0.370402688 -0.820979296 -2.06263691
#> units24  0.0506789 -1.34992127  0.085726429 -1.537396026 -1.26508071
#> units25  1.1959373  1.25405537  0.775186737  1.059809949 -0.71242521
#> units26  0.1686554  0.89224565 -0.943511216 -0.427406673 -0.07668162
#> units27 -0.3783570 -0.75775110 -0.545935202 -1.533029152 -0.73864864
#> units28 -1.6279852  0.69039161  0.377691639 -0.009423297 -0.31409250
#>            Trocterr    Zoraspin
#> units1   0.46963399  0.23019514
#> units2   0.58415245  0.60518294
#> units3   0.61101269 -0.22953141
#> units4   0.86254723  1.30829205
#> units5   0.91145758  0.99648545
#> units6   2.05141550  2.12169933
#> units7   1.19415154  0.95643791
#> units8  -0.03840288  0.11651861
#> units9  -1.41260325 -1.04966908
#> units10 -0.65729430 -1.91229775
#> units11 -1.08212043 -1.07515715
#> units12 -0.54204839 -1.52124732
#> units13  2.20627171  1.55703396
#> units14  0.95345027  1.50581646
#> units15 -0.14805972  0.12588715
#> units16 -0.08738422  0.22401234
#> units17 -0.22025833  0.01284372
#> units18 -0.14588418 -0.89177686
#> units19 -0.23897373  0.14103193
#> units20 -0.25080092 -0.10288163
#> units21 -0.36737073  0.01245841
#> units22 -0.87413764 -0.18068313
#> units23 -0.96508823 -3.31135756
#> units24 -1.05520423 -1.98770070
#> units25  1.08240349  0.92177439
#> units26 -0.66934895 -1.31676518
#> units27 -2.11056028 -1.41322772
#> units28 -0.50115196 -0.98119923