Simulates new data from a given cord object

Simulates new data from a given cord object

Usage

# S3 method for cord
simulate(object, nsim = 1, seed = NULL, newdata = object$obj$data, ...)

Arguments

object

is a cord object, e.g. from output of cord

nsim

Number of simulations, defaults to 1. If nsim > 1, the simulated data will be appended.

seed

Random number seed, defaults to a random seed number.

newdata

A data frame in which to look for X covariates with which to simulate.

...

not used Defaults to the X covariates in the fitted model.

Examples

abund = spider$abund

spider_mod_ssdm = stackedsdm(abund,~1, data = spider$x, ncores=2)
spid_lv_ssdm = cord(spider_mod_ssdm)
simulate(spid_lv_ssdm, nsim=2)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        5        0        1        0        0        0        1       71
#>  [2,]        0        0        0        0        0        1        1       35
#>  [3,]        3        1        5        0        4        0        4        0
#>  [4,]        0        0        0        0        0       13        5        0
#>  [5,]        2       15        0        0        0        0        0        4
#>  [6,]        0        2        0        0        0        8        4        0
#>  [7,]        3        2        0        1        0        0        1        4
#>  [8,]        1        0        1        0        0        0        0      108
#>  [9,]       12        0        0        0        0        4        0       51
#> [10,]        0        0        0        0        0        1        0       34
#> [11,]        0        2        0        0        0        4        4        0
#> [12,]        2        0        0        0        0        0        0       52
#> [13,]        8        5        5        1        0       14        0       39
#> [14,]        0       11        0        0        0        7        3        2
#> [15,]       10        8        0        3        0       29        0        1
#> [16,]        0        0        0        0        0        0        5        0
#> [17,]       10        8        0        0        0        1        0       11
#> [18,]        0       39        1        0        0        0       35        2
#> [19,]        3        6        0        0        0        6        0       16
#> [20,]        0        0        0        0        4        0       11        3
#> [21,]        2       16        0        0        0       34        0        4
#> [22,]        0        9        0        4        0        1       23       14
#> [23,]        0        9        0        0        0        3       11        8
#> [24,]        0       16        0        1        0        3        1       16
#> [25,]        0        1        0        0        0        0        1        4
#> [26,]       16       10        2        0        0        1        0       29
#> [27,]        5       30        0        0        0       11        0        5
#> [28,]        3       21        0        2        0        1        0       14
#> [29,]        0       17        0        0        0        2        0       30
#> [30,]        6        5        0        0        0        0        0       24
#> [31,]        7       16        0        0        0        3        3      167
#> [32,]       16        2        0        1        0       23        2       59
#> [33,]        3        0        0        0        0        0        1        0
#> [34,]        0        5        0        0        0        0       21        0
#> [35,]       14        0       14        0       11       20        0       26
#> [36,]        3        3        1        0        0        0        2        2
#> [37,]        3       17        0        0        0       13        6       68
#> [38,]        4        0       12        0        0        0        1        4
#> [39,]        6        0       28        0        3        0        2        0
#> [40,]        1        0        0        0        0        0        0        1
#> [41,]        4        5        0        0        0        1       10        5
#> [42,]       14        0        0        0        0        2        0        0
#> [43,]        3        7        0        0        0        9        4       48
#> [44,]        0        5        0        0        0        1        0        0
#> [45,]        2        1        0        0        0        0       16        9
#> [46,]        1        1        0        2        2        0        1        0
#> [47,]       43        0       10        0        8        0        0       60
#> [48,]        8       28       26        0        0        6        0       37
#> [49,]        1        0        0        0        0        0       10        1
#> [50,]        3        1        3        0        0        0        0        4
#> [51,]       19       15        1        1        0        2        0       17
#> [52,]        0        0        2        0        0        0        1        0
#> [53,]        2        0        4        0        0        0       10        1
#> [54,]        8        0        0        0        0        0        0       29
#> [55,]        0        0        0        0        0        0        0        0
#> [56,]       44        1       37        0        7        0        0      124
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        1        9        6        0
#>  [2,]        0        0       25        0
#>  [3,]        0        0        5        0
#>  [4,]        0        8       27        2
#>  [5,]        7       73       34        1
#>  [6,]        0        4      115       44
#>  [7,]        2        0       17        3
#>  [8,]        3        0       20        4
#>  [9,]        0       39        8        0
#> [10,]        0        1       19        0
#> [11,]        1       11       56       14
#> [12,]        0        0        6        0
#> [13,]       27       22       36        5
#> [14,]       72       74       84       14
#> [15,]       22       20       70        3
#> [16,]        0        0       20        2
#> [17,]       17       10       37        6
#> [18,]       99       11      191       16
#> [19,]        3        1       25        9
#> [20,]        0        0       16        1
#> [21,]       21       18       76       33
#> [22,]        4       12       94       37
#> [23,]      116      158       80       41
#> [24,]       36       98       94       17
#> [25,]        6        0       21        0
#> [26,]        0        1       27        1
#> [27,]       32       84       49        8
#> [28,]       27        4       24        4
#> [29,]        0       33       63        7
#> [30,]        0        0       35        4
#> [31,]       11       14       26       19
#> [32,]        7       27       39        2
#> [33,]        0        0       10        1
#> [34,]        0        0       50        6
#> [35,]        0        0        7        2
#> [36,]        0        8       44        0
#> [37,]        8       11       79        5
#> [38,]        0        0        2        0
#> [39,]        0        0        1        0
#> [40,]        0        0        1        0
#> [41,]       19        0       48        1
#> [42,]        0        0        6        0
#> [43,]        9        8      112        3
#> [44,]       10       17       41       18
#> [45,]        0        0        4        0
#> [46,]        0        0       14        4
#> [47,]        0        0        0        0
#> [48,]      157       16       59        8
#> [49,]        0        0        7        2
#> [50,]        0        0       11        1
#> [51,]       13       30       11        3
#> [52,]        0        0        2        2
#> [53,]        0        0        1        0
#> [54,]        7        2       20        1
#> [55,]        0       79       39        1
#> [56,]        0        8        0        0

# \donttest{
# using mvabund
library(mvabund) #for manyglm
abund=mvabund(abund)
spider_mod = manyglm(abund~1)
spid_lv = cord(spider_mod)
simulate(spid_lv)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        2        0        0        0        0        0        4        2
#>  [2,]        0        4        0        0        0        0        9        0
#>  [3,]        0        6        0        3        0       28       11       13
#>  [4,]        0        1        2        0        0        0        2        3
#>  [5,]       71        7       13        0       10        0        0      117
#>  [6,]        0        1        0        0        0        0       17        0
#>  [7,]        4        0       20        0        0        0        2        6
#>  [8,]        0        0        0        0        0        0        6        0
#>  [9,]        0        7        0        0        0        1        4       13
#> [10,]       18        1        2        0        0        0        1        1
#> [11,]        0        0        0        0        1        0        2        0
#> [12,]        0        1        0        0        0        0        7        0
#> [13,]       33        0        9        0        0        0        0       11
#> [14,]        1        2        0        3        0        0       32        0
#> [15,]       31        0        1        0        0        0        0       89
#> [16,]        0        0        0        0        0        0        7        0
#> [17,]        1        2        0        0        0        2       33        5
#> [18,]        0        4        0        2        0       28        0       10
#> [19,]        1        0        0        0        2        0        0        0
#> [20,]        5        0        2        0        0        0        9        1
#> [21,]        0        0        0        0        0        0       22        0
#> [22,]        2       10       14        1        0        1        0       65
#> [23,]        9       24        0        4        0       27        5       66
#> [24,]       18       15        1        3        0        7        0      210
#> [25,]        0        0        0        0        0        0        4        1
#> [26,]       77        0       43        0       44        0        0        5
#> [27,]        2       11        0        0        0        0        0       17
#> [28,]        1        8        0        0        0       15        0       31
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        1        0        1        6
#>  [2,]        0        0       56        4
#>  [3,]       14       73      157       33
#>  [4,]        0        0        3        0
#>  [5,]        0        7        2        0
#>  [6,]        1        0       27        7
#>  [7,]        0        0        2        0
#>  [8,]        0        0        3        0
#>  [9,]        3       15       40        8
#> [10,]        0        0        3        0
#> [11,]        2        0       39        3
#> [12,]        1        9       22        0
#> [13,]        0        0        0        0
#> [14,]        0        0       24        0
#> [15,]        0        0        1        0
#> [16,]        1        0        5        3
#> [17,]        4       22       39       10
#> [18,]        3       24       15        9
#> [19,]        0        0        2        0
#> [20,]        0        0        6        0
#> [21,]        0        0        5        0
#> [22,]        1        0       34        3
#> [23,]      117       70      144       30
#> [24,]       77      574       44        0
#> [25,]        0        0       11        1
#> [26,]        0        0        0        0
#> [27,]        0       25       12        1
#> [28,]        3        0       42       10

spider_mod_X = manyglm(abund ~ soil.dry + bare.sand, data=spider$x)
spid_lv_X = cord(spider_mod_X)
Xnew = spider$x[1:10,]
simulate(spid_lv_X, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        2        0        0        0        0        2       11        0
#>  [2,]        0        8        0        2        0        0       34        1
#>  [3,]        7        5        0        0        0        0        0       15
#>  [4,]        3        0        1        0        0        0        0       12
#>  [5,]        0        0        0        0        0        0       21        1
#>  [6,]        0        3        0        0        0        0        0        1
#>  [7,]        0       20        0        0        0        5        0       27
#>  [8,]        4        2        1        0        0        0        4        0
#>  [9,]        0        0        0        0        0        0       10        0
#> [10,]        1        0        2        0        2        1        0        3
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        2        0       23        6
#>  [2,]        0        1       36        5
#>  [3,]        5       34       29        0
#>  [4,]        0        5       16        0
#>  [5,]       24        1       22        3
#>  [6,]        9        0       55       35
#>  [7,]        0        4       77        3
#>  [8,]        0       21       23        0
#>  [9,]        2        0       10        1
#> [10,]        2        0        5        1
simulate(spid_lv_X, nsim=2, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        0        0        0        0        0        0        3        0
#>  [2,]        1        3        0        0        0        2        1        1
#>  [3,]        5        6        1        1        0        0        0       29
#>  [4,]        1        4        0        2        0        1        2        0
#>  [5,]        1       13        0        4        0        7        0       78
#>  [6,]        0       10        0        1        0        2        0        0
#>  [7,]        1        1        0        2        0        0        2       12
#>  [8,]        2        2        0        0        0        0        3        0
#>  [9,]        1        0        1        0        0        0       49        3
#> [10,]        0        0        0        0        3        0        0        0
#> [11,]        7       18        0        0        0        0        1      120
#> [12,]        3        2        0        1        0       13        4       44
#> [13,]        0        0        0        0        0        0        2        0
#> [14,]       14        4        1        2        0        1        0       12
#> [15,]        7       48        0        2        0        1        9      141
#> [16,]        0       25        2        5        0        3        0        7
#> [17,]        2       11        0        0        0       49        1        1
#> [18,]       14        7        1        0        0       11        0       62
#> [19,]        5        2        1        0        0        2        2       25
#> [20,]        2        1        4        0        0        4        0       16
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        0       10        2
#>  [2,]       15       86       84        2
#>  [3,]       30       23       25        1
#>  [4,]        1       30       48        5
#>  [5,]       11      297       83        3
#>  [6,]       29       22       76       48
#>  [7,]        0        3       29       15
#>  [8,]        0        0        5        1
#>  [9,]        0        3       20        4
#> [10,]        0        0        2        1
#> [11,]       17       40       34        2
#> [12,]       24      195       34        0
#> [13,]        0        0        6        1
#> [14,]        7       92       55        1
#> [15,]       18      237       90        7
#> [16,]       28        3       38       46
#> [17,]       47       26       97       17
#> [18,]        6        5       34        5
#> [19,]        1       22       19        0
#> [20,]        3        0       16        4

spider_mod_X_ssdm = stackedsdm(abund, formula_X = ~. -bare.sand, data = spider$x, ncores=2)
spid_lv_X_ssdm = cord(spider_mod_X_ssdm)
simulate(spid_lv_X_ssdm, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       20        1        0        0        0        0        0       34
#>  [2,]        1        4        1        1        0       59        0        7
#>  [3,]        9       12        0        1        0       14        0       12
#>  [4,]        6        1        0        0        0        5        0       76
#>  [5,]        1        1        0        1        0       13        1        6
#>  [6,]        0        7        0        1        0        4        3       12
#>  [7,]        8        7        1        4        0        6        0       57
#>  [8,]        0        3        0        0        0        2       22        0
#>  [9,]        6        1        0        0        0        0        0       63
#> [10,]       12        4        5        0        0        0        0       33
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        2       22        9        1
#>  [2,]       13       70      112       13
#>  [3,]       45       42       61        7
#>  [4,]        1       47        6        0
#>  [5,]       70       17       54       15
#>  [6,]       28       19       56       27
#>  [7,]      154       49      240       25
#>  [8,]        2        0       27        0
#>  [9,]        2        1       11        0
#> [10,]        4        1       12        1
# }