Calculate residuals from a stackedsdm object

Calculate residuals from a stackedsdm object

Usage

# S3 method for stackedsdm
residuals(object, type = "dunnsmyth", seed = NULL, ...)

Arguments

object

An object of class stackedsdm;

type

Determined what type of residuals to calculate. The current options include Dunn-Smyth residuals (default; "dunnsmyth"), raw response residuals ("response") or probability integral transform residuals ("PIT");

seed

For Dunn-Smyth and PIT residuals applied to discrete responses, random jittering is added, and the seed can be used to seed to jittering.

...

not used

Value

A matrix of residuals

Details

Calculated the residuals from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2) 
residuals(fit0)
#>           Alopacce   Alopcune    Alopfabr    Arctlute    Arctperi    Auloalbi
#> units1   0.8983379  0.9041709 -0.56999295 -1.39969157 -0.62456389 -0.37323939
#> units2  -0.8372316 -0.9761631 -1.59575240 -0.29111777  0.61443115 -0.80419868
#> units3   0.5249704  1.2577158  0.56120810  1.50226762 -0.31490374 -0.12634885
#> units4  -1.5605606  0.1565074  0.07757104  0.08586428  1.30878897  1.46548323
#> units5  -0.2721706  0.9843718  0.34120268  0.38602394  0.61757282  0.32644842
#> units6  -1.0405415  0.2080171  0.53688426  1.14298716  1.35668511  1.33669216
#> units7   0.1651406 -0.1284695 -2.36084854  1.01814389 -1.75835103  1.21415866
#> units8  -1.6706865  1.6943469  0.96004640  0.63820148  0.01481739  0.97920469
#> units9  -1.8679822 -0.1104851 -0.13621044 -0.60202423  1.05925703  0.11098341
#> units10 -0.7592675 -0.7991692  0.01598948  0.15589862  0.19064289 -0.94735832
#> units11  0.8558290 -0.4534234  0.26398810  0.85181599 -0.70461192  0.57382315
#> units12  0.6647019  1.7851363 -0.52769053 -0.53471139 -0.08308029 -0.76807665
#> units13  1.1411492  0.1920334  1.38686004 -0.81241497 -0.58916793 -0.06699383
#> units14 -0.5932118 -0.4233479 -0.55768855 -1.42758184 -0.59354696 -1.84549342
#> units15 -1.1805022 -1.6297098  0.63643180 -0.27212474  1.19699185 -2.42525832
#> units16  1.6136922  1.5787254  0.02965853  0.59791606 -2.01059996  1.19872811
#> units17  1.1737098 -1.5686822  0.85063039 -1.57260334  0.66109636 -0.05643163
#> units18  0.4886750 -0.8465679 -2.57248617 -0.14653482  0.81056186 -0.49700160
#> units19 -1.1722918  0.3435511 -1.10592286 -2.27507701  0.16775250  2.31880549
#> units20 -0.5649938  0.6917253  0.78054707 -1.74600233 -0.73597549  0.78041015
#> units21  1.0597489 -0.2824927 -1.25419638 -1.92577059 -0.18435027  0.59867817
#> units22  1.7460436 -1.0602337  0.65830749  1.72343300  0.05975956  0.28403083
#> units23 -0.6778769 -1.1757172  0.27821379 -1.65326638 -0.73340737 -0.89125195
#> units24 -0.2557749 -0.8297043  1.86411988 -0.51038884  1.38982147  0.62537442
#> units25  1.7492470  0.3141516  1.50883181 -0.78829235 -1.97689656  0.25425105
#> units26 -1.2476440 -1.4964929 -1.44954926 -0.32187161  0.10141769 -0.26239221
#> units27  0.6727237 -2.3094023 -0.34854812  0.73087181  1.09128683 -0.99284636
#> units28 -0.7675579 -0.7503692 -0.11324195  0.36861235 -0.93651471 -0.13291436
#>            Pardlugu   Pardmont    Pardnigr     Pardpull   Trocterr    Zoraspin
#> units1  -1.59355215  0.5986904  1.46306095  1.556957007  2.4483486  1.05653441
#> units2  -0.93283501 -1.4072956 -0.40068127 -1.062251256 -0.3168619 -0.55939312
#> units3   1.92977204 -0.1306934  0.84985451  0.147464459  1.5083475 -0.71135501
#> units4   0.95188964 -1.4244534  0.71189293  0.738532316  1.4429791  1.96503056
#> units5   0.54919304 -1.0314731  0.66245832  1.301653109 -0.4654699 -0.31047972
#> units6  -1.47279357  0.8157322  1.51275423  1.491429920  0.3425193  1.33101405
#> units7   1.04725988  0.8874069 -0.00813052  0.247286516 -0.1788272 -0.62005302
#> units8   1.73153276  0.8811304  0.61263388  0.167522944  0.6162493  0.39072589
#> units9  -0.97087358 -0.4980432 -0.23503846 -0.639054796 -1.6310992 -0.57579655
#> units10  1.15086850  0.7074806 -0.36987094 -1.981563858 -1.5062785 -1.34866250
#> units11  1.20857461  1.2696395 -0.97846890 -1.091136319 -0.7551494 -0.40647846
#> units12  0.57528996  1.0715803 -0.61944382  0.042064821  0.2433190 -1.19129157
#> units13  1.28864683  0.7734708  0.57116240 -0.781937672 -0.2882403  1.15899614
#> units14  0.99978192 -0.5670073 -1.36741080 -1.207517335 -0.2804200 -0.63557917
#> units15 -0.48093621 -2.0102248 -2.49717816 -1.008660820 -0.8316148 -0.50819898
#> units16 -0.23697754  0.3336165  2.14800591  0.738175633  1.3491551  1.42533317
#> units17 -1.05851515 -0.6130733 -1.81354891 -1.122163265 -0.9422148 -0.77769207
#> units18 -0.63333947 -0.2364734 -1.68276642 -1.000221031 -0.6725550 -1.56183432
#> units19  0.12886673 -2.5334745  0.48300698  0.735129080  0.7970772  0.49294717
#> units20  1.18291777 -0.2825942 -1.50692415 -0.002846483  0.6511571  0.29816458
#> units21  0.09674785  1.2916199 -0.59185775  2.542390448 -0.1004669  0.05363279
#> units22 -3.35543253  0.7622333  0.60443784  1.222657538  0.9078418 -1.25866201
#> units23  1.80766475 -0.3095862 -0.16147756 -1.106983856 -0.3699875  0.59711724
#> units24 -0.81750381 -1.1365291 -0.65346174 -1.246571246 -1.5949950 -0.31336136
#> units25 -0.77681148  1.8503362  0.88648088 -0.570989471  0.1166272  1.26119341
#> units26 -0.77322278 -0.5827482  1.09197673  2.131520494 -0.2414853 -0.13135145
#> units27 -0.27814958 -1.1022458  1.11935018  0.015874215 -1.4191683  0.14863992
#> units28 -0.64760873  0.1792680 -0.15065995  0.659000365  0.3629157 -0.64456325

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
residuals(fit0)
#>            Alopacce    Alopcune   Alopfabr    Arctlute    Arctperi    Auloalbi
#> units1   1.68231674 -0.32059190 -0.2212519 -1.14714160  0.45611647  0.42662348
#> units2  -2.20835478 -0.11855575 -0.6871815 -0.75209377  0.11303574  1.50819703
#> units3   1.26753676 -1.15342061  2.0604895  0.99285030 -0.74038193  0.82092647
#> units4   0.38346771 -0.15568986 -0.1548784  0.67797270 -0.97957675  1.34298273
#> units5   0.98621258  0.36487873  0.6343300  1.07163460 -0.77310028  0.80390734
#> units6  -2.26498089  0.69575180 -0.9777010  2.02107762 -1.19870561  1.35175574
#> units7   0.96933633  0.37598177  1.0223724  2.01983659  1.63440236  1.10417168
#> units8  -0.04229012  1.02074560  0.3660902  0.06137865 -0.76341019  0.67043714
#> units9   1.43602981 -0.54688327  1.0740342  0.51564138  0.07869367 -0.19348955
#> units10  0.90480945  0.42308143  0.2955406 -2.17313736 -0.88940297 -0.27207284
#> units11  0.94854159 -1.10616273  2.3688818  0.29896110 -0.65614045 -0.01242445
#> units12  0.62969461  1.48006396 -0.6802923  0.40263642  0.81668845 -0.53031323
#> units13  0.50730653  0.50784447 -0.3554614  1.25023594  0.44388424  1.91458686
#> units14 -0.50717885  2.72378682 -0.4499987  0.79315056  0.57023075  0.43154254
#> units15 -0.39299647 -2.05817313 -0.2882937  0.24213437 -0.46990320 -1.47692997
#> units16 -0.20454445  0.80906103  0.6475553  0.37705069  1.89477051 -0.13869125
#> units17 -1.04097731 -1.30374934  0.5027426 -1.32552831 -0.88501543 -1.31465303
#> units18 -0.40183783 -0.56139408  1.1712948  0.26431751  1.18816042 -1.16182101
#> units19 -0.02886262 -0.46652470 -0.2914950 -0.71739764  0.06835706 -0.11654440
#> units20 -0.31730778 -0.16488932  0.9645346 -0.51791030  1.16196602 -1.63726141
#> units21  0.04503323 -0.42953204 -0.1670997 -1.32836553 -0.77456999 -0.10805523
#> units22  0.56080414 -0.62277580  1.9285788  0.83521327  0.54420983  0.02089983
#> units23  2.76500998  0.19787487 -0.4069928 -0.88823086  1.36653736 -0.80759399
#> units24  0.51360124 -0.84527405  0.8797063 -0.98280847  0.70828111  0.13510062
#> units25  1.45931023  3.01974941 -1.0647895 -0.90767402 -0.70914780  1.04185613
#> units26 -1.92493499 -0.09208997  0.9101185 -0.59522154  0.33182743 -0.76483281
#> units27  0.34204517 -0.02424077  0.6479226  0.17421543  1.53217125  0.10680516
#> units28  0.71182521  0.52193981 -0.9075091 -1.31888983 -0.14362117 -1.03275794
#>            Pardlugu    Pardmont    Pardnigr     Pardpull    Trocterr
#> units1  -2.18312326  1.30309364  0.49198214  0.843042915  0.46986315
#> units2  -0.27124366 -0.70103861  0.57661829  0.759294111  0.58923178
#> units3  -0.56843828  0.76964765  0.65520774  0.840876165  0.59838702
#> units4  -0.52163831  0.05123008  0.84951971  1.181641603  0.85216571
#> units5  -0.51005249 -0.42151650  1.75701462  1.073146829  0.91318458
#> units6  -3.15847725  0.60614900  1.57541989  1.626437472  2.04729645
#> units7  -0.52588756  0.79601585  1.45679648  1.238212944  1.18989163
#> units8   2.53594159 -0.45068382  0.04863147 -0.127765703 -0.02107034
#> units9  -1.77856054  0.70728012 -0.10031887 -0.146642755 -1.47447345
#> units10 -0.01579194  1.37149034 -2.20784327 -0.713594337 -0.73497801
#> units11 -0.80804663  1.73479323 -0.97499361 -0.100847597 -1.09558646
#> units12 -0.66519485  1.74541668 -0.23561565  0.268663383 -0.57196320
#> units13  0.30735356  0.98798542  1.82775879  2.072382328  2.21361597
#> units14  0.03854435  0.33739557  0.58717286  1.049621339  0.94430509
#> units15  0.44659771 -0.77032569 -0.46763907 -1.109846859 -0.16458055
#> units16  0.42479709 -1.00804090  0.03671877 -0.792066623 -0.07333294
#> units17 -0.20056835 -1.84619609 -0.80075068 -1.764274967 -0.20993094
#> units18  0.28502882 -1.53495034 -0.40297934 -0.493359448 -0.14691131
#> units19  0.90800351 -0.85833320 -0.20132732 -0.298830283 -0.23173140
#> units20  0.95639224 -1.04727383 -0.73315652 -0.586560821 -0.24059973
#> units21  1.16074780 -0.50688731 -0.62632155 -0.139296733 -0.37259350
#> units22 -1.18363988 -0.03387868 -0.41032312 -1.553331362 -0.79321680
#> units23  1.37732281  0.40103997 -0.78943820  0.158619200 -0.74541385
#> units24 -1.20201222  0.08557960 -0.40670437 -1.281235134 -1.05956716
#> units25  1.61638118  0.79807452  1.06313989 -0.596227641  1.09922372
#> units26 -1.67140484 -0.72501552  0.09965221 -0.661455116 -0.37107891
#> units27 -0.68190022 -0.42914463 -0.01192385 -1.743547789 -2.21335178
#> units28  0.31553881  0.39038644  0.11297666 -0.005734301 -0.53280248
#>            Zoraspin
#> units1   0.22256248
#> units2   0.62716970
#> units3  -0.12987717
#> units4   1.28940882
#> units5   0.99803036
#> units6   2.13371256
#> units7   0.97788150
#> units8   0.07802329
#> units9  -2.22462806
#> units10 -0.39055434
#> units11 -1.14628941
#> units12 -0.87148955
#> units13  1.53774115
#> units14  1.50697953
#> units15  0.17066987
#> units16  0.25836137
#> units17 -0.11990480
#> units18 -0.51026314
#> units19  0.08588384
#> units20 -0.12435793
#> units21 -0.08683034
#> units22 -2.45365918
#> units23 -1.23750045
#> units24 -1.16763665
#> units25  0.86807071
#> units26 -1.43812410
#> units27 -0.91603098
#> units28 -1.09877511