Calculate residuals from a stackedsdm object

Calculate residuals from a stackedsdm object

Usage

# S3 method for stackedsdm
residuals(object, type = "dunnsmyth", seed = NULL, ...)

Arguments

object

An object of class stackedsdm;

type

Determined what type of residuals to calculate. The current options include Dunn-Smyth residuals (default; "dunnsmyth"), raw response residuals ("response") or probability integral transform residuals ("PIT");

seed

For Dunn-Smyth and PIT residuals applied to discrete responses, random jittering is added, and the seed can be used to seed to jittering.

...

not used

Value

A matrix of residuals

Details

Calculated the residuals from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2) 
residuals(fit0)
#>            Alopacce    Alopcune    Alopfabr   Arctlute    Arctperi    Auloalbi
#> units1   0.89969225  0.84303948 -0.89246456 -1.1867599 -1.72446843 -0.42679519
#> units2  -2.17445709 -0.93569667 -0.05135082 -0.7787639  0.97583318 -0.82489515
#> units3   0.40908862  1.28056994  0.53281607  1.7571806 -0.38788654 -0.10300463
#> units4  -1.45024401  0.11299847 -0.25110622  0.4722875  1.29619186  1.49444505
#> units5  -0.35480293  0.98105657 -0.81949961  0.6030235  0.15232875  0.26896145
#> units6  -0.62682490  0.19766347  1.53283744  1.2292527  0.62354472  1.36300990
#> units7   0.08904677 -0.17655410 -0.40305017  1.0283155 -0.19218037  1.22894104
#> units8  -0.66950096  1.61026923  1.09065765 -0.6234782  0.09825430  1.03109803
#> units9  -1.55163818 -0.05619147 -0.64555663 -0.5516516  0.21022191 -0.14450962
#> units10 -1.04463603 -1.51348243 -0.38551871 -0.7955979 -0.48028576 -0.87381477
#> units11  0.92985683 -0.38884332  0.54069543  1.1509373 -0.54307825  1.44898184
#> units12  0.74807158  1.80402092 -0.39147996 -1.3659933  0.45328038 -0.74555508
#> units13  0.86341651  0.18644170  1.34228442 -0.5689420  1.75747741 -0.09047412
#> units14 -0.20249612 -0.46691399 -0.59664768 -1.2034573 -0.64696280 -1.75148593
#> units15 -0.19782652 -2.34713850 -0.81043563  1.1074608  0.13695520 -0.16150083
#> units16 -1.43984770  1.36663687 -0.65937252  1.7695355  1.38634996  1.06387820
#> units17  1.46767468 -1.33417049 -0.49873571 -0.4700431 -1.61529663 -0.84122565
#> units18  0.46106964 -0.89927974  0.98560816  0.9523477 -0.59363659 -1.36516989
#> units19 -0.39077243  0.05879708  0.28298849  1.7317646  0.75074173 -0.94600927
#> units20 -2.00065806  0.68151116 -1.58247722  1.7637739 -1.04455471  0.22673701
#> units21 -0.46154314 -0.09971215  1.22914344 -0.6834911 -1.73493693  0.71071665
#> units22  1.84629115  0.14501107  0.62124655  0.3840026  0.01129154  0.04138896
#> units23 -0.66456313 -1.84522764  0.29394821  0.2387859 -0.47864200  1.40769346
#> units24 -0.24638086 -0.34526262  1.86568949  0.7131561  1.55711350 -1.28134224
#> units25  1.77494756  0.23111755  1.52994113  1.2197572  0.51232599  0.28951347
#> units26 -1.31704433  0.24325631 -1.42040060 -0.1780324  0.08553886 -1.16256183
#> units27  0.66702996 -0.30906999 -0.32844685 -0.0838475  0.85310197 -2.17244849
#> units28 -0.78115624 -0.95663411 -0.13635522  0.2728638 -0.84563745  0.22397194
#>            Pardlugu   Pardmont     Pardnigr    Pardpull     Trocterr
#> units1  -0.38709080  0.6042089  1.469012115  1.56255067  2.442951351
#> units2  -0.56555259 -1.5142585 -0.389245834 -1.07380512 -0.323883888
#> units3   2.05700308 -0.1629322  0.855682239  0.14599145  1.535579277
#> units4   1.08949763 -1.4297064  0.684357235  0.74073811  1.447781607
#> units5   0.32251632 -1.0459080  0.660369281  1.29598151 -0.459191374
#> units6  -1.64950610  0.8002861  1.527263312  1.46181189  0.339546826
#> units7   0.70137932  0.8951377 -0.005664027  0.24162686 -0.176883973
#> units8   1.74146079  0.8735517  0.668914199  0.18536052  0.620990073
#> units9  -0.43138430 -0.5076528 -0.084144198 -0.25505581 -1.568046841
#> units10 -0.72975079  0.7163432 -1.221800178 -1.17945119 -1.675966648
#> units11 -0.76639884  1.2608398 -0.347725393 -1.35105598 -0.791172679
#> units12 -0.56750823  1.0651054 -0.597553291  0.04543431  0.266804060
#> units13  1.27469440  0.7770377  0.573706458 -0.77141690 -0.274799373
#> units14  0.74143780 -0.5768564 -1.365385364 -1.20522575 -0.286391966
#> units15 -0.44818178 -0.9214257 -1.258468616 -0.43121394 -0.837235627
#> units16 -0.26166061  0.7413367  2.100449325  0.86056556  1.340596000
#> units17 -0.99119698 -1.0137075 -0.956956928 -0.27335374 -0.968543447
#> units18 -0.71191133  0.1902716 -1.270447816 -1.15100803 -0.669685910
#> units19  0.24275989 -0.1741095  0.463066411 -0.83147586  0.803797398
#> units20  1.22157636 -2.3774221 -1.241468521 -0.73774764  0.719631436
#> units21  0.05859838  1.2989921 -0.618451344  2.50273556 -0.083793229
#> units22 -0.15507861  0.3825483 -0.094538611 -1.47862819  0.656142571
#> units23  2.26643302 -0.3172301  0.476607896  0.10300354 -0.001876724
#> units24  0.56343608 -1.1286664 -1.558929904  0.61580986 -1.742467589
#> units25 -1.15591869  1.9265018  0.882410595 -0.78744752  0.145826467
#> units26 -0.70875024  0.1392954 -1.637260396 -1.00161596  0.387501214
#> units27 -1.76915719 -1.1141568  0.080733942  1.22331000 -0.559562430
#> units28 -0.40854316  0.1752579 -0.435220002  0.29452817  0.226076172
#>            Zoraspin
#> units1   1.07431762
#> units2  -0.59563390
#> units3  -0.58914776
#> units4   1.98546987
#> units5  -0.29202764
#> units6   1.34046699
#> units7  -0.61823277
#> units8   0.14282874
#> units9  -0.72402311
#> units10 -0.36193047
#> units11 -0.81269781
#> units12 -0.38856010
#> units13  1.19015298
#> units14 -0.61874204
#> units15 -0.53971467
#> units16  1.44170040
#> units17 -0.72534046
#> units18 -1.14329782
#> units19  0.64124906
#> units20  0.56836928
#> units21 -0.08627374
#> units22  0.14566576
#> units23  0.86578691
#> units24 -1.08315749
#> units25  1.26841160
#> units26 -0.25500391
#> units27 -0.77876136
#> units28  0.27934491

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
residuals(fit0)
#>            Alopacce   Alopcune    Alopfabr    Arctlute   Arctperi    Auloalbi
#> units1   0.73201180  0.3384430 -1.40164619  0.35615378  0.5803397  0.43740041
#> units2  -0.20052503 -0.1129811 -0.37308518 -1.03121549 -0.9029922  1.50766500
#> units3   0.87990437  0.1942110  1.30544624  0.94462774 -1.1041851  0.78248022
#> units4   1.46664082  1.5716850 -0.57158618  0.67526328  0.5831165  1.34426498
#> units5   0.28421847 -0.1416695 -0.27773442  0.96706948  0.8344467  0.80823538
#> units6  -1.87751351  0.9379356 -1.32877610  2.01078990 -0.3904796  1.34414653
#> units7   0.73679749  2.3952682 -0.06460806  2.04486688  0.7424365  1.09485814
#> units8  -1.28962955  0.2964231 -0.19726699 -0.02675464 -1.1112286  0.66483154
#> units9   1.68037084 -0.5475759 -1.61403462 -0.05792422 -0.2224450 -1.91425623
#> units10  0.18097911  0.1652557  1.14873691 -2.90253604 -1.5233171  0.05774528
#> units11  1.30284953  0.1284765  1.71381324 -1.87906662 -0.1771539  0.01121812
#> units12  0.29499821  0.5516457  0.81239319  0.07709195  0.7542477 -0.95900878
#> units13 -0.60982935  0.3056515  0.66721160  1.30821087 -0.2759767  1.93873823
#> units14 -0.10094087  1.0374042  0.70180771  0.63259520  1.4591651  0.48487533
#> units15 -0.09307468 -1.6910434 -1.13537911 -1.69971540 -0.8164822 -0.62009000
#> units16  0.06177745 -0.9986040 -0.41357967  0.07046859  0.2287378 -0.85148096
#> units17 -1.03213193 -1.3145611  0.55691442  0.53520761  0.9934212 -1.14614163
#> units18 -1.14816842 -0.7353681  1.05727453 -0.13540728  0.1308240 -0.41572693
#> units19 -1.22624741  0.4594784 -0.54996107  0.06678059  0.8867120 -0.95229821
#> units20 -0.56667630  0.3952263 -0.66513080 -1.51665822  0.5452638 -0.50434472
#> units21 -0.39257794 -1.0510031 -0.03393877  0.24738122  1.5085177 -1.30688439
#> units22  1.31527783 -0.6561146 -0.21298784  0.29140853  0.5959460  0.11648588
#> units23  0.12652663 -1.7228152  1.15629577 -1.30026816  1.3921982  0.11375460
#> units24 -0.68354681 -0.6722710  0.04385111 -0.18634789  0.6964723  0.20328869
#> units25  0.56669290  1.5001086  0.33622435 -0.84873802 -1.0323720  1.08393671
#> units26 -1.14991150 -0.5946497 -1.61697843 -0.48998667  0.3367243 -0.60238665
#> units27 -0.53574035 -0.8911729 -0.04811623 -1.44452373  1.7890216 -0.40029458
#> units28  0.41712604  0.2023096 -1.08226343 -1.01597697 -0.2845561 -1.65384484
#>            Pardlugu     Pardmont    Pardnigr    Pardpull    Trocterr
#> units1  -0.82973139  1.301800056  0.51522915  0.83879595  0.47523940
#> units2  -0.31770178 -0.507776906  0.59438069  0.76732566  0.59372876
#> units3  -0.54103595  0.761435413  0.65003743  0.84519556  0.59794433
#> units4  -0.30661716  0.031813515  0.84702897  1.17899992  0.85795467
#> units5  -0.43010402 -0.417048690  1.75688143  1.07302989  0.91593364
#> units6  -0.42096006  0.593658986  1.56991350  1.62966369  2.03624864
#> units7  -0.47671762  0.785627545  1.45188575  1.23813512  1.19487751
#> units8   2.52042251 -0.402373058  0.05027899 -0.15387941 -0.03243422
#> units9  -1.84264519  0.702313140 -0.21184837 -0.13225020 -1.45606844
#> units10  0.48137571  1.393815283 -0.77271827 -0.49565735 -0.51726063
#> units11 -1.62294732  1.730969055 -0.69808716 -0.22938271 -1.09458830
#> units12 -1.50985393  1.740409825 -0.11074617  0.28825941 -0.53678064
#> units13  0.10792153  0.972437471  1.83148078  2.08029070  2.20855444
#> units14  0.07367867  0.338670873  0.57423430  1.04671155  0.94904975
#> units15  0.44727085 -3.362033546 -0.57899794 -1.47353105 -0.16570229
#> units16  0.39182874 -0.872336363 -0.05055453 -1.75642737 -0.06931060
#> units17 -0.14780150 -0.754884123 -0.45492540 -0.57886007 -0.20755375
#> units18  0.32726192 -1.089104945 -2.96007436 -1.01181505 -0.15754217
#> units19  0.92162569 -1.164657308 -0.10385848 -0.70799576 -0.22993402
#> units20  0.98117152 -0.955607018 -0.35681665 -2.20143028 -0.24609430
#> units21  1.13771503 -0.644931234 -1.42625180 -0.10426820 -0.36157568
#> units22  0.51805091 -0.040958168 -0.56282949 -0.44383157 -0.69771570
#> units23  1.58492811  0.332676077 -0.54312241 -1.70704411 -0.95849509
#> units24 -1.46049982  0.004591269 -0.37222392 -1.11124989 -1.93004457
#> units25  1.01629675  0.770798002  1.09367542  0.08047232  1.11977730
#> units26 -0.36431996 -1.195345589 -0.12218431 -1.91704749 -0.51159998
#> units27  0.25961327 -0.387446988 -0.02299066 -0.56066189 -1.82109452
#> units28  0.41465520  0.385771101 -1.84887898 -0.90812711 -0.52467188
#>            Zoraspin
#> units1   0.25598779
#> units2   0.63310760
#> units3  -0.27454885
#> units4   1.28902865
#> units5   0.99991779
#> units6   2.15078424
#> units7   0.96027155
#> units8   0.10504437
#> units9  -1.67402194
#> units10 -2.27211412
#> units11 -1.08298587
#> units12 -1.65361827
#> units13  1.55990844
#> units14  1.49922501
#> units15  0.14850961
#> units16  0.23310331
#> units17 -0.04052453
#> units18 -1.37807426
#> units19  0.11881767
#> units20 -0.11421160
#> units21  0.04099016
#> units22 -0.49225458
#> units23 -0.55667451
#> units24 -0.81831423
#> units25  0.93927589
#> units26 -1.43392890
#> units27 -0.47259436
#> units28 -0.20038141