Calculate residuals from a stackedsdm object

Calculate residuals from a stackedsdm object

Usage

# S3 method for stackedsdm
residuals(object, type = "dunnsmyth", seed = NULL, ...)

Arguments

object

An object of class stackedsdm;

type

Determined what type of residuals to calculate. The current options include Dunn-Smyth residuals (default; "dunnsmyth"), raw response residuals ("response") or probability integral transform residuals ("PIT");

seed

For Dunn-Smyth and PIT residuals applied to discrete responses, random jittering is added, and the seed can be used to seed to jittering.

...

not used

Value

A matrix of residuals

Details

Calculated the residuals from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2) 
residuals(fit0)
#>           Alopacce    Alopcune    Alopfabr    Arctlute    Arctperi    Auloalbi
#> units1   0.8906432  0.86460667 -1.11454484 -0.54820806 -0.33504609 -0.57474044
#> units2  -0.3210869 -0.80630827 -0.63697594 -0.20404641 -1.78259068 -0.79928737
#> units3   0.4788192  1.25963080  0.90786755  1.37855188 -0.31806738 -0.13727207
#> units4  -1.6027955  0.18622766 -1.42940827  0.55256171  0.49260042  1.47835842
#> units5  -0.7456409  0.98034835 -0.99521913  0.72345389  1.38998269  0.36551317
#> units6  -0.5730707  0.18978805 -0.71429689  1.21213536 -1.06394462  1.53108967
#> units7   0.5440305 -0.21333026 -1.12268416  1.04463238 -0.63591014  1.18141814
#> units8   0.8304587  1.67533768 -1.04491161  0.95313935 -0.81758132  0.99025422
#> units9  -1.7617402 -0.40036998 -0.46411796  0.50072512 -0.96519829 -0.58560510
#> units10 -0.9826930 -2.05281004 -0.33306146  0.16231494  0.48253350  0.48211888
#> units11  0.8525420 -0.42099865  0.25013770  0.41729729 -0.20066006  0.52587850
#> units12  0.6378517  1.79523189 -1.15853451 -0.41258655  0.02540018 -1.02174200
#> units13  1.1469034  0.19419777  1.19077250 -0.80286798  2.17768900 -0.05879494
#> units14 -1.1171225 -0.49117662  0.15779641 -1.44333530  0.12083635 -1.79157793
#> units15  0.3579667 -1.41608364 -0.01999707  0.58454194  0.66125650 -0.14026300
#> units16 -0.4281039  1.28285017 -1.53122744 -0.62017940 -2.32796688  0.32997121
#> units17 -0.6255433 -2.09505834  0.63952797  0.98323316  2.32380800 -0.15480088
#> units18 -0.5163215 -0.76878208 -0.55212545 -0.46840156  0.27704174 -0.45542101
#> units19 -0.8397358  0.20314575 -1.08365219 -0.71797299  0.12523182 -0.76416707
#> units20 -1.2289984  1.10119408  1.12203122  0.51722925  1.18568762 -0.95933242
#> units21  0.8490193  0.01378390 -0.28484486  0.88486807  0.94873625 -0.65586807
#> units22  1.7527425 -0.18928118  0.61434008  0.44408783 -0.35115952 -0.46760785
#> units23 -0.6612150 -0.43285751  0.25321795  0.40711910 -0.49720429 -0.56619865
#> units24 -0.1715066  0.07120448  1.86891061  0.08341252  1.48403152  0.26323373
#> units25  1.7339468  0.35100654  1.50021482  0.15689476  1.42299659  0.78757209
#> units26 -1.2822911 -0.48501938 -1.45095668  0.27671126  0.11555326  1.54801802
#> units27  0.7191211  0.07811514 -0.33233404 -0.11824652  1.23081140  0.16732140
#> units28 -0.8120151 -0.88801463 -0.17800023 -0.16307932 -1.06183717 -1.48972781
#>            Pardlugu   Pardmont      Pardnigr    Pardpull    Trocterr   Zoraspin
#> units1   0.63123862  0.5982089  1.4648447079  1.58768412  2.42633589  1.3402792
#> units2  -0.85172158 -1.5847598 -0.3979552033 -1.07754034 -0.32970576 -0.5974501
#> units3   1.26442812 -0.1452943  0.8299233734  0.17140865  1.50472102 -0.8510447
#> units4   0.76599698 -1.4690412  0.6944294966  0.74442894  1.45550734  1.9352554
#> units5  -0.10423861 -1.2049154  0.6626342749  1.29905203 -0.47113817 -0.3006401
#> units6  -2.20227857  0.8108446  1.5383009415  1.44411054  0.34493368  1.3406535
#> units7   0.36306202  0.8833151  0.0009161788  0.24925148 -0.18482608 -0.6264045
#> units8   1.74380854  1.2896878  0.5545768002  0.15402838  0.65080172  0.2291492
#> units9  -0.83564581 -0.5058852 -0.2615742222 -0.75521710 -1.63846458 -0.7832267
#> units10 -2.01060322  0.7252249 -2.9455522773  0.51172004 -1.50194716 -0.7053619
#> units11 -1.46158595  1.2633050 -1.1075608648 -1.02533902 -0.80649485 -0.2840616
#> units12 -1.38452137  1.0662077 -0.3792635131 -0.14186062  0.21685418 -0.3816897
#> units13  1.84109642  0.7705084  0.5731393468 -0.77758020 -0.28550004  1.1915478
#> units14  1.00256318 -0.5669387 -1.3808996763 -1.20162610 -0.28528682 -0.6217912
#> units15 -0.40671179 -0.5818516 -1.3418487969 -0.80996742 -0.83494893 -0.5035622
#> units16 -0.42638259  0.7883976  1.7750418360  0.17488344  1.34906831  1.2312314
#> units17 -1.07982939 -0.8537699 -1.8194928127 -0.36796274 -0.96528381 -0.6103345
#> units18 -0.72451127 -0.1285083 -0.7262516609 -2.08652373 -0.66295544 -1.5422799
#> units19  0.23372640 -0.6445031  0.7804109682 -0.66029715  0.79779061  0.3690572
#> units20  1.24796150 -0.3521064  0.0221641344  0.04150662  0.65203398  0.1475935
#> units21  0.06545209  1.3826343 -0.3835084044  2.36889183 -0.06626097  0.0077021
#> units22  0.15333432  0.3337648  2.0865530425  1.03578367  0.45139737  0.6070647
#> units23  2.17802069 -0.3009345  0.7260057184 -1.91813767 -0.38078189  0.8811943
#> units24  0.25776002 -1.0495295  0.6011626962  0.11773776 -1.28483648 -0.1031685
#> units25 -0.90243392  1.8788532  0.9178865419 -1.03009866  0.09682338  1.5030434
#> units26 -0.43373917 -1.9448241 -0.5903020226  2.33348137  0.45217849  0.5716442
#> units27 -1.29708066 -1.1322186 -0.2049964927  0.19688212 -1.19067722  0.7801740
#> units28 -0.59953741  0.1531032  1.2609586728 -0.44635945  0.69147523  1.3576098

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
residuals(fit0)
#>            Alopacce    Alopcune    Alopfabr    Arctlute    Arctperi    Auloalbi
#> units1   0.63933829  0.47205106 -0.05742870  0.24525245  0.07546295  0.47013763
#> units2  -1.35839968  0.04993850 -0.92391394 -0.67097964 -1.09805673  1.51313413
#> units3   1.36522160 -0.51127081  2.53163064  0.93244648 -0.15848852  0.82750082
#> units4   1.26860767 -0.09871615  0.88246754  0.66719971 -0.86372476  1.34021065
#> units5   2.53043852  1.63699227  0.02698368  1.06429770 -0.41840452  0.79002018
#> units6  -2.21927660  0.39021179 -2.77699150  2.04437212 -1.60886584  1.29251060
#> units7   1.85914309 -0.67114404 -0.30159223  2.01083474  0.22513565  1.08368303
#> units8  -0.16277518 -0.31891404  0.19206221 -0.25054918 -0.14192912  0.72043875
#> units9   0.20225662 -0.18902845 -0.75667245 -0.55425843 -0.74828924 -0.48298753
#> units10 -0.60614424 -0.43664024 -0.02512726 -0.86811805 -0.94245933 -0.17650882
#> units11  0.35150353  0.58411358  1.30976587  0.30461971 -0.38855918  0.05012167
#> units12  0.69792306  1.15517498  1.04806136 -0.03132944  0.67040495 -1.91028217
#> units13  0.19386078  1.22869265  0.64427049  1.17771325  0.42548522  1.93150568
#> units14 -2.09572832  1.40575675  0.41189003  0.69070710 -1.54494332  0.47014652
#> units15 -0.06394519 -1.77880863 -0.04336619  0.17790628  0.75427642 -1.35491010
#> units16 -1.32083417  1.73827490  1.15217696 -0.17441146  0.82165299 -0.98700138
#> units17 -1.41236564 -2.26936386 -0.20939339 -0.75600959  1.67489003 -1.04130483
#> units18 -0.00165947  1.39500886  0.97426308  0.47035587 -1.01813508 -0.15094111
#> units19 -0.06990364 -0.04004767  0.08769638  0.35128275  1.39593986 -0.59279507
#> units20  0.05187384 -0.07542525  0.19046086  0.35903447  0.96506033 -0.30629048
#> units21 -0.09228023  0.12524815 -0.28954131 -0.66758481 -0.48576058 -0.84750110
#> units22 -0.77676694 -0.18041298  0.05198890  0.27236619  0.53247403 -0.74891547
#> units23  0.53686424  0.12647852 -0.35370113 -0.52590527  1.30719015  0.24331475
#> units24 -0.24741611  0.43548849 -0.36381097  0.61710002  0.78070707  0.06209281
#> units25  1.84337331  2.17904206  0.81072797 -1.40034073 -1.13230385  0.96623085
#> units26  1.27384708 -1.68709026 -1.35906036 -0.71558937  0.32851704 -1.00845838
#> units27  0.90904182 -0.60294719  1.12710221 -0.99908516  1.60792458 -1.63562200
#> units28 -0.09971718 -2.59821833  1.56726640 -1.70266347 -0.30611759  0.23436818
#>            Pardlugu    Pardmont    Pardnigr    Pardpull    Trocterr
#> units1  -0.61317390  1.29851099  0.49742171  0.83713415  0.47089737
#> units2  -0.57891073 -0.62602203  0.58086140  0.76633045  0.58379929
#> units3  -0.35465141  0.76598254  0.64129273  0.84511067  0.60582003
#> units4  -0.58100806  0.05312566  0.84927330  1.18200115  0.85688886
#> units5  -0.39420237 -0.47464476  1.75712508  1.07122367  0.91352882
#> units6  -1.17182930  0.60376359  1.56308880  1.62278999  2.03951504
#> units7  -0.39906688  0.79866387  1.45724950  1.24060376  1.19241775
#> units8   2.52341600 -0.38496671 -0.04031980 -0.24333399 -0.03506072
#> units9  -0.62545356  0.69138175 -0.12366193 -0.14421749 -1.34059347
#> units10  0.74419264  1.38861903 -0.76702107 -0.44371177 -0.74206646
#> units11 -1.83176712  1.73343068 -0.94568788 -0.09300502 -1.07637086
#> units12 -2.08641808  1.74002337 -0.07259128  0.29662497 -0.56878555
#> units13  0.46466831  0.97601511  1.83045712  2.07490003  2.21620513
#> units14  0.07735881  0.34927673  0.59070735  1.04563224  0.94439186
#> units15  0.36719276 -1.20132493 -0.55441494 -0.30636421 -0.14561562
#> units16  0.38850997 -1.76372529 -0.01609190 -0.67459870 -0.08564880
#> units17 -0.13736091 -1.06107292 -0.95398269 -1.78699848 -0.21528128
#> units18  0.33830200 -1.17379704 -2.21978754 -0.68414044 -0.14818912
#> units19  0.89147383 -1.37776181 -0.11177719 -0.30839283 -0.23207849
#> units20  0.94163068 -1.26998730 -0.61524822 -0.40068794 -0.23471998
#> units21  1.12532611 -0.70878527 -0.76973670 -0.11757763 -0.35357066
#> units22 -1.00497271 -0.17099822 -1.29910868  0.23951724 -0.83799142
#> units23  1.50262352  0.36084562 -0.02703393 -0.03773769 -0.87444449
#> units24 -1.65929459  0.11274437 -0.01995034 -0.83587066 -1.36538680
#> units25  1.32872358  0.80359343  1.03189042 -1.09749311  1.13932588
#> units26  0.44573271 -1.45426033 -0.70898083 -0.60360650 -0.54691705
#> units27  0.19003133 -0.46548605 -0.81381826 -0.39339931 -1.43742617
#> units28 -0.90134398  0.42867646 -0.45156732 -0.91814261 -0.55216820
#>             Zoraspin
#> units1   0.229291928
#> units2   0.645364071
#> units3  -0.357235653
#> units4   1.281483309
#> units5   1.014443708
#> units6   2.123879026
#> units7   0.990498967
#> units8   0.084991862
#> units9  -0.625937087
#> units10 -0.603875146
#> units11 -1.850028347
#> units12 -0.662776495
#> units13  1.552157620
#> units14  1.494856971
#> units15  0.173703367
#> units16  0.247064486
#> units17 -0.059597222
#> units18 -0.705315086
#> units19  0.144930335
#> units20 -0.003679344
#> units21 -0.103699333
#> units22 -0.597209036
#> units23 -1.110773340
#> units24 -1.531624039
#> units25  0.929880351
#> units26 -0.231329479
#> units27 -0.257459541
#> units28 -1.544548846