Calculate residuals from a stackedsdm object

Calculate residuals from a stackedsdm object

Usage

# S3 method for stackedsdm
residuals(object, type = "dunnsmyth", seed = NULL, ...)

Arguments

object

An object of class stackedsdm;

type

Determined what type of residuals to calculate. The current options include Dunn-Smyth residuals (default; "dunnsmyth"), raw response residuals ("response") or probability integral transform residuals ("PIT");

seed

For Dunn-Smyth and PIT residuals applied to discrete responses, random jittering is added, and the seed can be used to seed to jittering.

...

not used

Value

A matrix of residuals

Details

Calculated the residuals from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2) 
residuals(fit0)
#>            Alopacce     Alopcune      Alopfabr    Arctlute     Arctperi
#> units1   0.92655314  0.926562961 -1.5187668008 -0.92057234 -0.674880449
#> units2   0.18671574 -0.819426260 -0.1508746227 -1.54932388 -0.570913328
#> units3   0.53687652  1.267312064  0.5621595164  1.44743121  0.090053565
#> units4  -1.61211300  0.135936973 -2.0253588225 -0.02015660 -0.230389057
#> units5  -0.85702028  1.017689679  0.4026451764  0.77574757  0.541186801
#> units6   0.64825294  0.161129746  0.3170918663  1.24301020  0.104551694
#> units7   0.10757886 -0.148178056 -0.1989489015  1.01889984  0.211802734
#> units8  -0.09960146  1.671738102 -0.7065634193 -0.05349393  0.756323771
#> units9  -1.31289777 -0.470441872 -0.2025842962 -1.56116580  1.673923846
#> units10 -0.97216344 -1.756158179  0.0070466614 -0.28096452 -0.849633612
#> units11  0.87693403 -0.271882123  0.2909423116 -2.22387016  0.733862405
#> units12  0.67774802  1.815927373 -0.9080141298 -0.82940928  0.422285598
#> units13  0.83121781  0.187263721  1.7835147527 -0.75713023 -0.992339686
#> units14 -1.65988623 -0.537340419  1.1795602149 -1.23774643  0.491175365
#> units15  0.32143424 -1.847420991 -0.3345558802 -0.56176737 -0.012652922
#> units16 -0.34290692  1.631866285  0.8654104069  0.02256920 -2.759318669
#> units17  1.47658041 -1.608725615  1.2581411988  0.91626026 -2.033909843
#> units18 -0.62780501 -1.098154826  0.0380706989 -0.50072840  0.309126631
#> units19  0.32353618  0.024156556 -0.3499669830  1.82796291 -0.874233772
#> units20 -0.86142093  0.811201544 -0.0007369668 -1.51631401  0.209235041
#> units21 -0.25293757 -0.024266580  1.3175403979 -1.23256244 -0.797652493
#> units22  1.98094164 -0.773280829  0.6779158256 -1.08754016 -0.289670555
#> units23 -0.70403244  0.116639889  0.2908031033 -1.04019870 -0.618751182
#> units24 -0.28060200 -1.387047811  1.8609192252 -0.50115935  1.268750462
#> units25  1.75266246 -0.005318714  1.5372571256 -0.25717556  0.308260392
#> units26 -1.26689857 -0.222496953 -1.3839438339 -2.02875796 -0.009717426
#> units27  0.65748030  0.531929049 -0.3257968946  0.99046569  1.055388634
#> units28 -0.81486338 -1.699477203 -0.1248119349 -1.02237027 -1.121147452
#>            Auloalbi     Pardlugu   Pardmont     Pardnigr     Pardpull
#> units1  -0.38786682 -1.994093315  0.6112552  1.420170800  1.585821783
#> units2  -0.79978870 -0.642837201 -1.7253097 -0.426666545 -1.065683920
#> units3  -0.18567570  1.858347887 -0.1626969  0.838221349  0.177408323
#> units4   1.50366859  1.319909037 -1.4406375  0.687183665  0.751597396
#> units5   0.30383138  0.002031487 -1.1863471  0.660741692  1.289554566
#> units6   1.21308437 -2.946473579  0.8455124  1.549588991  1.453892293
#> units7   1.13233955  0.329168718  0.8761269 -0.006253238  0.249094422
#> units8   1.00507743  1.743235380  1.2635846  0.471198463  0.073900088
#> units9  -2.52787545 -2.089671705 -0.4761915 -0.353322209 -0.327300018
#> units10 -0.59437439 -2.051629527  0.6861396 -0.012768060 -0.526812510
#> units11  1.47114209 -1.591135205  1.2739413 -1.399915365 -0.804571133
#> units12 -1.52514224 -0.682667322  1.0786541 -0.590961953  0.043660678
#> units13 -0.03989521  1.017541832  0.8087999  0.590060206 -0.769531256
#> units14 -1.76166193  0.732890225 -0.5907754 -1.376124448 -1.214224390
#> units15 -0.81888306 -0.461853803 -1.9230280 -0.918532578 -1.432198724
#> units16  1.43086353 -0.380513770  0.7763228  1.840802592 -0.875473894
#> units17 -0.51981175 -1.079177588 -1.1836553 -1.994923424 -1.983678091
#> units18 -0.67426141 -0.513078519  0.3551815 -0.690506140  0.335354794
#> units19  1.30250503  0.174310168  0.5026151  0.819715355 -0.216476784
#> units20 -2.16687884  1.281782040 -0.8260298 -0.394245033 -0.896028989
#> units21 -0.43641818  0.085241193  1.7839596 -1.241184527  2.616862801
#> units22  0.72155644  0.691192478  0.3649273 -0.309294298  0.025639929
#> units23  1.53557596  2.398512577 -0.2511448  0.046595525  0.005305597
#> units24  0.89443374 -0.721582274 -1.0948795  0.619609954 -1.087444770
#> units25  0.74843215 -0.852309634  1.9841596  0.951910615 -1.335705317
#> units26 -0.31655610  0.088290790  0.2962548 -0.568310811  1.581829512
#> units27 -0.37110828 -0.461209479 -1.4027073  0.337982311 -0.260818214
#> units28 -0.85021367 -1.922781879  0.2841344 -0.450215460  0.801101974
#>            Trocterr   Zoraspin
#> units1   2.44183511  1.2833576
#> units2  -0.33425825 -0.5982250
#> units3   1.51055777 -0.7159753
#> units4   1.45435124  1.9666506
#> units5  -0.47124426 -0.3154142
#> units6   0.33696354  1.3195073
#> units7  -0.17229520 -0.5825429
#> units8   0.61503453  0.1312760
#> units9  -1.75687741 -1.4072782
#> units10 -1.78941774 -0.2563467
#> units11 -0.76608195 -1.1369852
#> units12  0.25253519 -1.8709964
#> units13 -0.28141095  1.2104384
#> units14 -0.27179696 -0.6207090
#> units15 -0.84266509 -0.5989424
#> units16  1.31875678  1.2967008
#> units17 -0.94133127 -0.4854970
#> units18 -0.67717424 -2.2340507
#> units19  0.80240809  0.5853675
#> units20  0.64796088  0.2860955
#> units21 -0.09090116 -0.2586711
#> units22  0.84042887  0.1214064
#> units23 -0.02595169 -0.8494288
#> units24 -1.31202840  0.2764693
#> units25  0.17292823  1.4625437
#> units26  0.08404882 -0.9726367
#> units27 -1.12748107  0.2799408
#> units28  0.38309521 -0.1263507

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
residuals(fit0)
#>            Alopacce    Alopcune      Alopfabr    Arctlute     Arctperi
#> units1   0.54059264  0.14859597 -0.0399052422 -0.35570838 -2.056887434
#> units2  -0.97657268 -0.38047296  0.1146993293 -0.91750824 -1.432663877
#> units3   2.29520020 -0.25000635  1.2299754910  1.00829876  0.009199622
#> units4   2.18743004  0.99347727 -0.1749069700  0.64992925  0.579991156
#> units5   0.44143798 -0.64019858 -0.3807767715  0.95463600 -1.677928222
#> units6  -1.85683533  0.77141745 -1.4996449408  1.96567342 -0.412461451
#> units7   0.46050766 -0.67548567  0.6064599611  2.00845442 -0.394505041
#> units8  -0.47264925 -0.68802952 -1.2097845018 -1.40460835  0.866341589
#> units9   0.67201249  0.37252159  1.1859079076 -0.71391231  1.152879714
#> units10  0.39172704 -0.25969087  0.6550665647  0.74503136 -1.022021056
#> units11  2.18488266 -0.27684610  2.2750998350 -1.19328663  1.213270039
#> units12  1.97701609 -0.20469343 -1.9400113928 -0.04359224  0.455121051
#> units13  0.67008601  0.23079929  0.2228305921  1.19129544 -0.435067019
#> units14 -0.61041457 -0.76916212  0.0759690666  0.70469627  0.238993318
#> units15  0.05505446 -1.70624280 -0.2938402342 -2.00896150 -0.248842054
#> units16 -0.49632889  1.02767165  0.2418072704 -0.33804984 -0.412151437
#> units17 -0.17535411 -1.30899839 -0.2868377009 -0.35288939  0.970783240
#> units18 -1.25747679 -0.39450068  1.1713997059 -0.68392165 -0.727314697
#> units19 -1.31518050  0.07414419 -0.6605814473 -1.50353196 -1.181217557
#> units20 -0.23453844  1.10205528  0.5479833056 -1.05455970 -0.240189755
#> units21 -1.35161567  0.27078266 -0.0006069754  0.13425383 -0.788542121
#> units22  1.30737963  0.34676697  0.5467710444 -0.22464127  0.530200751
#> units23  0.08739961 -1.32549537  0.2206758040 -0.56318285  1.282277248
#> units24  1.65478142  0.66391917 -0.1375929262 -0.99375449  0.718774819
#> units25  0.47889556  1.99827865 -0.3618907472  0.21698923 -0.991755931
#> units26 -0.53232605  0.51387193 -1.5358577852 -1.37948906  0.318592569
#> units27  0.66044657 -1.75831023  1.0995844340 -1.79176430  1.769327911
#> units28  0.49412599 -1.22032603 -1.1263543138  0.47832380 -0.498062184
#>            Auloalbi    Pardlugu    Pardmont    Pardnigr    Pardpull    Trocterr
#> units1   0.49985096 -1.25630117  1.30808662  0.50752484  0.83762272  0.47276871
#> units2   1.51457008 -0.30827375 -0.60583482  0.57898301  0.76022234  0.58506983
#> units3   0.79326030 -0.41959438  0.77486036  0.66207677  0.84376819  0.59849541
#> units4   1.34480813 -0.45591654  0.03275508  0.85371085  1.17969079  0.85225193
#> units5   0.78779789 -0.40242635 -0.36862004  1.75748674  1.07530752  0.90955673
#> units6   1.28268162  0.20088985  0.58182126  1.58004877  1.63385942  2.03646314
#> units7   1.11331351 -0.47686467  0.78486911  1.45429371  1.23667933  1.19059408
#> units8   0.66965170  2.53932643 -0.41799244 -0.01469444 -0.20172923 -0.02431748
#> units9  -0.65380687 -0.73674687  0.69447305 -0.13637309 -0.17941221 -1.40240719
#> units10  0.13759755 -1.19516155  1.39041044 -0.74825168  0.44584762 -0.53563310
#> units11  0.09514646 -0.84102242  1.72745103 -1.12805467 -0.10783532 -1.11958661
#> units12 -1.04422839 -2.40694601  1.73848421 -0.23969157  0.29209971 -0.54584888
#> units13  1.93194749  0.18928589  0.97980288  1.81908468  2.08043455  2.21343450
#> units14  0.51348753  0.08407083  0.36040531  0.58398741  1.04731181  0.95354637
#> units15 -0.43170736  0.44563652 -1.68318085 -0.93593068 -0.40299945 -0.14899323
#> units16 -1.43073631  0.36564100 -0.87018900  0.01183984 -1.86663522 -0.06639075
#> units17 -0.83674192 -0.20528565 -1.43227972 -0.66670503 -1.11815467 -0.20051124
#> units18 -0.88782704  0.25887327 -0.77508553 -0.28949545 -0.63428539 -0.15797454
#> units19 -0.61438622  0.90747742 -2.17929559 -0.08572928 -0.34781418 -0.23959895
#> units20 -1.56886024  0.96468190 -0.85872359 -0.25680740 -2.00099327 -0.24202666
#> units21 -1.47779343  1.13242480 -0.69855072 -0.62255296 -0.12271507 -0.37279255
#> units22  0.28666955 -0.21869356 -0.06977897 -0.29241861 -0.23917218 -0.68506590
#> units23 -0.56858413  1.47663012  0.33645009 -0.49146756 -1.11077592 -0.95517107
#> units24 -0.88362464 -0.22887682  0.12737426 -0.88406639 -0.40151705 -1.39150783
#> units25  0.95612573  1.40802638  0.82432346  1.06592762 -1.02393044  1.11105788
#> units26  0.44061073  0.26056341 -0.70436717  0.03109321  0.51882475 -0.47928344
#> units27 -1.04490014 -0.13413683 -0.41885028 -1.85380626 -0.62131243 -2.55107295
#> units28 -0.93924949  0.25026510  0.42140215 -1.48741660 -0.03262264 -0.41558535
#>            Zoraspin
#> units1   0.17687933
#> units2   0.60090956
#> units3  -0.21201170
#> units4   1.30843042
#> units5   1.02160482
#> units6   2.12444137
#> units7   0.99081825
#> units8   0.11605163
#> units9  -3.50788070
#> units10 -0.23414191
#> units11 -1.16578311
#> units12 -0.64472433
#> units13  1.55693092
#> units14  1.50579398
#> units15  0.11799907
#> units16  0.20597267
#> units17 -0.09804606
#> units18 -2.02136626
#> units19  0.04975208
#> units20  0.01357058
#> units21 -0.10135779
#> units22 -0.55939318
#> units23 -0.54984178
#> units24 -0.22297100
#> units25  0.89395077
#> units26 -0.91078762
#> units27 -1.02580248
#> units28 -1.43841640