Calculate residuals from a stackedsdm object

Calculate residuals from a stackedsdm object

Usage

# S3 method for stackedsdm
residuals(object, type = "dunnsmyth", seed = NULL, ...)

Arguments

object

An object of class stackedsdm;

type

Determined what type of residuals to calculate. The current options include Dunn-Smyth residuals (default; "dunnsmyth"), raw response residuals ("response") or probability integral transform residuals ("PIT");

seed

For Dunn-Smyth and PIT residuals applied to discrete responses, random jittering is added, and the seed can be used to seed to jittering.

...

not used

Value

A matrix of residuals

Details

Calculated the residuals from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2) 
residuals(fit0)
#>            Alopacce    Alopcune    Alopfabr    Arctlute   Arctperi    Auloalbi
#> units1   0.84676395  0.90113447 -0.60660144  0.15442680 -0.6311492 -0.62360740
#> units2  -0.80591386 -0.90472309  0.28382635 -1.09652669  2.7051703 -0.79258430
#> units3   0.53716968  1.25368356  0.73320080  1.51361387 -0.3309628 -0.15235196
#> units4  -1.56962127  0.18754882 -2.58323935  0.31953099  1.1630902  1.45042873
#> units5  -0.70007835  0.98237346  0.45162688  0.40049171 -1.6871373  0.23480007
#> units6  -0.60162141  0.08204508 -0.21648113  1.25389243  0.7615577  1.34188517
#> units7   0.36801414 -0.16992164 -0.03210456  1.07116900 -1.0861311  1.11917089
#> units8  -0.93997030  1.60616294  1.19209686  0.52261975  0.3880035  0.83515524
#> units9  -1.43981252 -0.49724220 -0.78355308  0.17591727  0.8650652 -0.50575128
#> units10 -0.82210122 -0.59294303 -0.25549161 -0.79428348 -0.3915386  0.25838852
#> units11  0.83029558 -0.27992850  0.33738096 -1.21552115 -0.5308967  1.06606521
#> units12  0.74126066  1.87958212 -1.65682064 -0.64846500  0.4267813 -0.74845115
#> units13  0.66189067  0.19097304  1.37202293 -0.82095276 -0.1155773 -0.08102301
#> units14 -0.12428711 -0.49407052 -0.43014544 -1.14900037 -0.6078006 -1.71908761
#> units15 -1.42665142 -1.62556832  1.47471821 -1.84034112  1.6192901 -0.97799988
#> units16  0.03836433  1.28028221  0.52346124  0.17898767  0.1903424  0.28472672
#> units17  0.61470968 -1.49254594  1.06649342  0.94987251  1.1452556 -1.84668864
#> units18 -2.99553019 -1.04457821 -0.30727629  0.60683392  0.1854106 -1.52041432
#> units19  1.47588554  0.23312224 -1.21051746  1.14777232 -0.1881938 -0.39372275
#> units20  0.46196853  0.89344266  0.12023512 -0.51104778 -0.6398224 -0.21288709
#> units21  0.99266321  0.16923131  0.30186560 -0.04853513 -0.2052426  0.21825730
#> units22  1.74921496 -0.71209329  0.61017559 -1.84854001 -0.3371181  0.47840319
#> units23 -0.66530306 -0.66986003  0.22765969 -1.07468897 -0.5993605 -1.37092530
#> units24 -0.18159553 -1.07773200  1.88618062 -0.94346708  1.3202356 -0.39004366
#> units25  1.77367701  0.43308449  1.48175391  0.68294069  0.3186063  0.26741492
#> units26 -1.29591520  0.14531344 -1.44386296 -0.83607019  0.1531407  0.36723046
#> units27  0.66102149 -1.03575240 -0.35697241 -0.22932974  0.9881757 -0.22635542
#> units28 -0.80409719 -0.72226129 -0.16656504 -0.54452761 -0.3709130  0.75625524
#>            Pardlugu    Pardmont     Pardnigr     Pardpull    Trocterr
#> units1  -0.15529979  0.61162906  1.388567774  1.580238645  2.43554057
#> units2  -0.74059598 -1.56539604 -0.394992875 -1.065985150 -0.33452021
#> units3   1.32450281 -0.14625283  0.847925882  0.147322828  1.50602885
#> units4   0.99418481 -1.41388288  0.715584584  0.748173860  1.45524225
#> units5   0.05696165 -0.98776657  0.663419641  1.305130860 -0.46411210
#> units6  -2.46656211  0.79082736  1.509057535  1.509600172  0.34794234
#> units7   0.67644082  0.88177118  0.001694762  0.241591034 -0.17404911
#> units8   1.73334998  1.14886093  0.749165801  0.836116548  0.61702647
#> units9   1.05510660 -0.49303435 -0.225385269 -0.296598391 -1.64092213
#> units10 -1.28544027  0.72152031 -0.089990886 -0.625627849 -1.62939799
#> units11  0.16008074  1.26199170 -0.348386039 -1.207344511 -0.91321762
#> units12  0.09982693  1.07779068 -0.347851496 -0.008974913  0.18519985
#> units13  1.42321372  0.81342008  0.588209944 -0.783199881 -0.27820064
#> units14  0.83218978 -0.62361557 -1.370200529 -1.203454186 -0.27737316
#> units15 -0.30531527 -1.96310455 -1.525919763 -0.914965315 -0.85293974
#> units16 -0.34609385  1.26191171  2.160604304  0.379924008  1.34661108
#> units17 -0.76133229 -1.81839424 -1.212556216 -1.046400861 -0.95727795
#> units18 -0.59152347 -0.04500548 -0.858840601 -0.783645808 -0.69396322
#> units19  0.14857238 -2.76876012  1.088228690 -0.994971191  0.78600749
#> units20  1.24765086  0.02444553 -0.031964786 -0.656284892  0.66770767
#> units21  0.06174067  1.22370247  0.126814017  2.586227799 -0.13931405
#> units22 -0.96333122  0.58206577 -0.330051706  0.701922779  1.02603911
#> units23  2.03192490 -0.22853509 -0.525335452  1.005356980 -0.08017796
#> units24 -0.53459813 -1.19476455  0.260915924  0.451479704 -1.95271802
#> units25 -1.28505093  1.90803762  0.964923910 -0.311800427  0.15784024
#> units26 -0.21827439 -0.51633915 -0.061548333 -1.734712297  0.45113334
#> units27 -0.84923222 -1.07313568 -0.433837627  1.265191408 -2.34110636
#> units28 -0.72362752  0.31798783  0.935715456  1.661593609  0.68401794
#>            Zoraspin
#> units1   1.20270292
#> units2  -0.64381557
#> units3  -0.68176021
#> units4   1.93188045
#> units5  -0.32977002
#> units6   1.29676287
#> units7  -0.59755860
#> units8   0.32678568
#> units9  -0.93120131
#> units10 -0.01442898
#> units11 -1.69465063
#> units12 -0.96918643
#> units13  1.20027315
#> units14 -0.61093435
#> units15 -0.46768975
#> units16  1.16411525
#> units17 -0.53751792
#> units18 -2.28768233
#> units19  0.42576996
#> units20  0.50473153
#> units21 -0.01702470
#> units22 -1.33452255
#> units23 -0.07235338
#> units24  0.53606500
#> units25  1.29568163
#> units26 -0.64525976
#> units27 -0.25685202
#> units28  0.31167168

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
residuals(fit0)
#>             Alopacce    Alopcune    Alopfabr     Arctlute    Arctperi
#> units1   1.092284826  0.17215280 -1.28437483 -1.671999373  0.69875790
#> units2  -1.765234638 -1.20957574  0.23761368 -0.369366852  0.05653940
#> units3   1.172792414 -0.66632813  1.46957671  0.921227361  0.67599721
#> units4   1.025075066 -1.26402183 -0.93140561  0.603600093  1.49302548
#> units5   2.416192603  0.28867107 -0.18740381  0.966922153  0.71904378
#> units6  -2.332430782  1.26854437 -1.46990497  2.077890590 -2.79520233
#> units7   0.547025681 -1.07429026 -1.31411328  2.068926515  1.18046291
#> units8  -1.784682057 -0.45589944  0.85432398 -0.580803966 -1.23653684
#> units9   0.164067610  0.78412837 -1.04650524  0.197864574  0.04533819
#> units10 -0.119758270  1.06168194  1.29736494  0.258358131 -1.45416278
#> units11  0.996861080 -0.04575339  1.51832978 -0.842509217 -0.68458393
#> units12  0.775420808 -0.93683514 -0.87625862 -0.121790967  0.44264081
#> units13  1.128804658  0.18127252  0.81392701  1.262413035 -0.66889709
#> units14 -0.124847791  0.79314749 -1.40767628  0.795600423 -1.36509529
#> units15  0.006266484 -1.33860105 -0.04661852 -0.168995469  1.22306377
#> units16 -1.271087026 -0.79608516 -0.03849232  0.006541949  0.30341108
#> units17 -1.625953413 -1.51751182  0.64329162  0.376468065 -0.56925213
#> units18 -1.172542830  0.86286196 -1.04877709 -0.507414789 -0.13451888
#> units19 -1.699116975  1.09322224 -1.59653761 -0.196095369  0.09123978
#> units20 -0.522534418 -0.07828422  1.00376420 -0.177512524 -1.18613673
#> units21 -0.257354192  0.88847912  0.24822430 -0.373885890  0.96309522
#> units22  0.328581076 -0.21365139 -0.96830754 -0.953173481  0.60494435
#> units23 -0.523151994  0.31521957 -0.61987914 -1.118210641  1.33489289
#> units24 -0.068986012 -0.52856140  0.82202143 -0.123824562  0.73910941
#> units25  0.280665820  1.56337339  0.49754263  0.676170607 -0.78115784
#> units26  0.219524843 -1.24428276  0.17704969 -1.162720842  0.33440117
#> units27  0.312785045 -1.43798238  0.65374666  0.656910994  1.66985327
#> units28 -0.055549347  0.23784659 -0.36074712  0.650662452 -0.14578664
#>            Auloalbi    Pardlugu    Pardmont    Pardnigr    Pardpull    Trocterr
#> units1   0.49127046 -1.12612323  1.30497694  0.51096600  0.84054776  0.47778733
#> units2   1.49732483 -0.44918225 -0.56557818  0.58700684  0.76884435  0.58594724
#> units3   0.81213275 -0.25859775  0.76654498  0.64915513  0.84464145  0.60350197
#> units4   1.35242610 -0.49950495  0.01986302  0.84955625  1.17756006  0.86165169
#> units5   0.81824607 -0.41762853 -0.36547825  1.75754146  1.07629490  0.91222296
#> units6   1.29133838 -0.30247578  0.60701039  1.56834752  1.63573704  2.04254333
#> units7   1.09501408 -0.38123293  0.79432699  1.45132810  1.23655651  1.19417230
#> units8   0.67975073  2.53515362 -0.47115406 -0.04816686 -0.20263645 -0.02497191
#> units9  -1.92836621 -1.48650473  0.70273527 -0.12895532 -0.22233778 -1.31328640
#> units10 -0.99348909 -0.43515886  1.37748022  0.11161866 -0.64954335 -0.75788900
#> units11 -0.06176787 -1.55919723  1.72711102 -0.38938276 -0.20006212 -1.13464700
#> units12 -0.35076601 -1.37610295  1.74177471 -0.09313934  0.29193461 -0.53674211
#> units13  1.90432874  0.11940141  0.97174216  1.82835460  2.07720649  2.21385171
#> units14  0.48990912  0.07024112  0.34198442  0.58123597  1.04745497  0.94398161
#> units15 -0.22878656  0.38599624 -1.99160333 -0.75934096 -1.12569336 -0.16625665
#> units16 -0.65217256  0.38465635 -0.96390392 -0.04240769 -0.35971676 -0.07257776
#> units17 -1.10798772 -0.07742627 -0.88310474 -0.87566189 -0.79095181 -0.20335296
#> units18 -0.36774508  0.29410941 -0.78672102 -1.25581276 -0.72374283 -0.16220929
#> units19 -0.99276713  0.87026225 -2.72840600 -0.08009601 -0.88900939 -0.24582606
#> units20 -0.26986761  0.95430690 -1.49959411 -0.89943040 -0.47669033 -0.23525012
#> units21 -0.53075442  1.15105825 -0.66476459 -1.54830204 -0.10512698 -0.35476091
#> units22 -0.19486808 -0.64727475 -0.01511828 -1.34934917  0.04050169 -0.85721580
#> units23  0.13708151  1.62413699  0.39499647 -1.13502808 -0.49005951 -0.94663692
#> units24 -0.93182081  0.15830051  0.02679960  0.14343189  0.16318302 -1.69848073
#> units25  0.94264656  1.10688792  0.81905196  1.09237823  0.04805092  1.11365184
#> units26  0.30572810  1.05612874 -0.66283818  0.01427232 -0.35272909 -0.57938222
#> units27 -0.87868845 -0.88307732 -0.52378449 -0.57650169 -1.49910055 -1.57885345
#> units28 -0.91058381  0.19943272  0.44461105 -2.12411278 -1.87464444 -0.58485384
#>            Zoraspin
#> units1   0.19559296
#> units2   0.59393381
#> units3  -0.23305194
#> units4   1.29502986
#> units5   1.01072224
#> units6   2.12284707
#> units7   0.95432447
#> units8   0.15845410
#> units9  -0.66586490
#> units10 -0.40394091
#> units11 -2.09683663
#> units12 -2.19871410
#> units13  1.55191200
#> units14  1.51443809
#> units15  0.06338783
#> units16  0.27912554
#> units17 -0.03646092
#> units18 -0.75771886
#> units19  0.07209813
#> units20 -0.09335670
#> units21  0.02334781
#> units22 -0.79562413
#> units23 -2.13032299
#> units24 -1.03292910
#> units25  0.91824403
#> units26 -0.62995488
#> units27 -0.63008756
#> units28 -0.19985497