Calculate residuals from a stackedsdm object

Calculate residuals from a stackedsdm object

Usage

# S3 method for stackedsdm
residuals(object, type = "dunnsmyth", seed = NULL, ...)

Arguments

object

An object of class stackedsdm;

type

Determined what type of residuals to calculate. The current options include Dunn-Smyth residuals (default; "dunnsmyth"), raw response residuals ("response") or probability integral transform residuals ("PIT");

seed

For Dunn-Smyth and PIT residuals applied to discrete responses, random jittering is added, and the seed can be used to seed to jittering.

...

not used

Value

A matrix of residuals

Details

Calculated the residuals from stackedsdm object.

Author(s)

Francis K.C. Hui <francis.hui@anu.edu.au>.

Examples

X <- spider$x
abund <- spider$abund

# Example 1: Simple example
myfamily <- "negative.binomial"
# Example 1: Funkier example where Species are assumed to have different distributions
# Fit models including all covariates are linear terms, but exclude for bare sand
fit0 <- stackedsdm(abund, formula_X = ~. -bare.sand, data = X, family = myfamily, ncores=2) 
residuals(fit0)
#>           Alopacce    Alopcune    Alopfabr    Arctlute     Arctperi    Auloalbi
#> units1   0.8444547  0.88917197 -0.56257711 -0.98852559  0.313521442 -0.40363854
#> units2  -0.8459242 -0.87047202 -0.06313992 -1.30298097 -0.502899469 -0.82590076
#> units3   0.4403751  1.26866843  0.73187864  1.33029850 -0.214209153 -0.20234028
#> units4  -1.2998946  0.16461441 -0.17639060  0.36572150  0.181337545  1.51567641
#> units5  -0.3722187  1.00178668 -0.66148199  0.53499940 -0.973965389  0.25318078
#> units6   1.2178799  0.19480197  0.18582389  1.13847907  0.287585352  1.41132079
#> units7   0.3146227 -0.18504972 -0.29286541  1.06580859  0.799611950  1.19083446
#> units8  -0.4559356  1.58340980 -0.62661855 -0.60422675  0.523784714  0.97154123
#> units9  -1.6783215 -0.49284110 -1.07810264 -0.03738564 -0.135106325 -0.60097894
#> units10 -0.7212328 -1.00601131 -0.36927975 -1.81503055 -0.177740280  0.28551593
#> units11  0.8122730 -0.06366139  0.70318205 -1.04341877  0.401857299  1.21847929
#> units12  0.6465360  1.83904253 -1.79692442 -0.40594389  0.024760297 -0.65124343
#> units13  0.9972428  0.18092903  1.72563992 -0.85039452 -0.921909382 -0.06561885
#> units14 -0.3060117 -0.50480099 -1.24518056 -1.46903550 -1.196888014 -1.76080423
#> units15  0.5906644 -1.28370664 -0.65811491 -0.65811759 -1.411998409 -0.61091055
#> units16  0.8262266  1.63905393  0.66177431  1.14771716  0.710625458 -0.22090078
#> units17 -0.6462910 -2.26027348  0.49047252 -0.70239590 -0.767883508 -0.41143088
#> units18  0.5338676 -0.79909110  1.51382642 -1.08450358  0.967595314 -0.46347024
#> units19  0.7534264  0.31881486  0.39163499  1.41946359 -0.103178329  0.46365227
#> units20  0.6466451  0.74522573 -0.11356970  2.00476725  1.218894907 -1.00993740
#> units21  0.1706872  0.01545507  0.45988043 -2.45426666 -1.171706976 -0.13289895
#> units22  1.7884615  0.84504922  0.63235117 -2.34790054  0.005937991 -0.14680854
#> units23 -0.6009841 -0.04788073  0.27539653  0.99501645 -0.721128159 -1.26167190
#> units24 -0.1835017 -0.33985886  1.93757004 -2.25586724  1.619104084 -2.12159113
#> units25  1.7575721  0.35164084  1.49166520  1.25791506 -0.447917611  0.76158909
#> units26 -1.3071711  0.56646721 -1.39687327  1.19059798 -0.008267210  0.05527394
#> units27  0.6756860 -1.13956631 -0.37753119 -1.92201675  1.373292640 -0.80850094
#> units28 -0.8120413  0.10074619 -0.15844489 -0.60386077 -0.836539420  0.22563189
#>           Pardlugu   Pardmont     Pardnigr    Pardpull    Trocterr    Zoraspin
#> units1  -0.6116494  0.5934639  1.408066903  1.57579101  2.43417215  1.08282903
#> units2  -0.9543090 -1.5047209 -0.432735256 -1.06495943 -0.33311330 -0.61490358
#> units3   1.5069833 -0.1454279  0.838821516  0.17111777  1.53550673 -0.85088957
#> units4   1.4336042 -1.5023279  0.688821076  0.73836861  1.44018217  1.93023551
#> units5   0.6518717 -1.1769895  0.655249855  1.29582665 -0.45768240 -0.31008623
#> units6  -1.4016726  0.8390272  1.553545102  1.50514711  0.34858703  1.30423275
#> units7   0.4158744  0.9150787 -0.007026821  0.25632182 -0.18350909 -0.59991435
#> units8   1.7648233  0.9407119  0.770494790  0.11107744  0.66054439  0.10550620
#> units9  -1.1801494 -0.4848943 -0.440503843 -0.25408153 -1.83446331 -0.37972785
#> units10 -0.1375594  0.6887601 -1.400572501  0.54377287 -1.94953433 -0.07572580
#> units11 -0.4675082  1.2725253 -0.899379294 -0.74654525 -0.92425501 -0.92392700
#> units12 -3.2709176  1.0658713 -0.534408275 -0.08747833  0.24741995 -0.40016768
#> units13  1.2998767  0.8098798  0.573623285 -0.77092038 -0.27741251  1.19372380
#> units14  0.7021638 -0.5888105 -1.374103554 -1.21959551 -0.27434659 -0.61267200
#> units15 -0.3303415 -1.0805540 -1.093184915 -0.32768073 -0.83443779 -0.67671222
#> units16 -0.2572956 -0.6106045  1.710611052  1.12861416  1.30815477  1.26803045
#> units17 -0.7830340 -0.3601811 -1.038951055 -0.85769615 -0.93038441 -0.48750746
#> units18 -0.6108507 -0.7023081 -1.132680792  0.23742738 -0.65036362 -1.67174942
#> units19  0.1580465 -1.0516263  0.695977510 -0.20663963  0.77447648  0.48308885
#> units20  1.2719863  1.3255713 -0.451268667  0.49943444  0.65802951  0.58345353
#> units21  0.1052263  1.0578484 -0.119974077  2.50272040 -0.07535812 -0.27454286
#> units22  1.0630185  0.7241631  1.171037614  0.70230081  0.92101756 -1.95405324
#> units23  1.9467983 -0.2752515 -0.367526659 -0.29309929 -0.12485667 -1.95431828
#> units24  1.0725140 -1.1242766 -1.026165549 -0.05588913 -1.92835278  0.09017762
#> units25 -0.8547044  1.9756932  0.781089403 -0.18080820  0.14412550  1.31134539
#> units26  0.7015977  0.3068376 -1.487361211  0.61393442 -0.04012573 -0.80517185
#> units27 -0.6610698 -1.0424017  0.217575625  0.29761489 -2.04094914 -1.46617566
#> units28 -0.1453819  0.2906818 -0.486707309 -1.28875433  0.27883878 -0.61374712

# Example 2: Funkier example where Species are assumed to have different distributions
abund[,1:3] <- (abund[,1:3]>0)*1 # First three columns for presence absence
myfamily <- c(rep(c("binomial"), 3),
              rep(c("negative.binomial"), (ncol(abund)-3)))
fit0 <- stackedsdm(abund, formula_X = ~ bare.sand, data = X, family = myfamily, ncores=2)
residuals(fit0)
#>            Alopacce    Alopcune    Alopfabr    Arctlute    Arctperi    Auloalbi
#> units1   0.30309169 -0.46724856 -0.07429489 -1.72999377 -1.11348062  0.44054738
#> units2  -0.48331789  1.79441955 -0.61425941  0.13466547 -0.61711772  1.50031188
#> units3   2.30596174  0.56397354  1.31891821  0.95185978 -0.18642773  0.81592191
#> units4   1.00894045  0.09560458  0.49827510  0.65586057  1.32614075  1.34004346
#> units5   0.57130882  1.29051019 -0.26570732  0.95064788  0.38270565  0.82278408
#> units6  -2.05856748  2.25436063 -1.35686148  1.96181979 -1.92587148  1.36435581
#> units7   0.95406760  0.79773642  0.10531345  2.01769248 -0.67244872  1.09618185
#> units8  -0.06618608 -0.93956088  0.78333767  0.04341422 -1.19022102  0.68061363
#> units9   0.42838245 -1.15558119 -0.28317520  0.24629677  1.03281937 -0.50237326
#> units10  0.46248865 -0.70888212  0.04936219 -0.50760265 -1.76997836 -0.75811882
#> units11  1.53757610  0.33802927  1.80625337  0.52758310 -0.09105471  0.06342050
#> units12  0.21027998  1.43869832  1.17019337 -1.58968796 -0.61348988 -2.10207401
#> units13  1.05151780  1.65398797  0.59198038  1.24567643  0.56893018  1.93984237
#> units14  0.09456915  0.95792579 -0.78644400  0.77326438 -1.11159378  0.43821494
#> units15  0.13948023 -1.63553787 -1.01878122  0.04577384 -0.69726113 -0.20526863
#> units16 -0.56969509 -0.29014785  0.39500586 -0.01155467 -0.41668456 -0.52928438
#> units17 -1.12432611 -1.33076358  0.18581029  0.02380823 -1.34582166 -0.36815916
#> units18 -1.23273167  1.33826297 -0.78797504 -0.31541327  0.29610403 -0.15796900
#> units19 -1.09645253  1.54732193 -0.91163947  0.44711722 -0.49027578 -2.22850361
#> units20 -0.55478676 -0.13997324 -0.77329213 -0.35002153  0.26651607 -1.13400509
#> units21 -0.41265371  1.39555489  1.06864465 -0.85124983  0.60231844 -0.71333146
#> units22  1.34721762  0.00811088 -0.06602349 -0.22906176  0.56684871 -0.36048953
#> units23  0.27448028  0.07051687  0.03451502  0.36612468  1.32543766  0.05673762
#> units24 -0.35539631 -0.12549077 -0.60024422 -0.80416204  0.76693437 -1.03474984
#> units25 -0.68494637  1.68045010 -0.25813451 -0.78477765 -0.76176431  1.11504271
#> units26 -1.43507959  0.61157671  2.69048604 -1.59211369  0.30679387 -0.18980604
#> units27 -0.27709305 -1.03093473 -0.05545055 -0.81604999  1.68676244 -0.52698158
#> units28  1.63633273  0.35297584 -0.95420585 -0.51776623 -0.23239867 -0.78540696
#>             Pardlugu    Pardmont    Pardnigr    Pardpull    Trocterr
#> units1  -1.290166578  1.30399917  0.49818717  0.83942127  0.47678703
#> units2  -0.461375963 -0.51381259  0.58716130  0.76032491  0.58461400
#> units3  -0.522767448  0.77523000  0.65309374  0.84087047  0.60662611
#> units4  -0.289913398  0.05456968  0.85011581  1.17771840  0.86283848
#> units5  -0.300624623 -0.44330391  1.75417909  1.07145435  0.91633099
#> units6  -0.544645825  0.57206964  1.57853747  1.62073151  2.03656767
#> units7  -0.432653533  0.79971069  1.45829455  1.23702169  1.19316707
#> units8   2.542118831 -0.35231602  0.02229763 -0.16637128 -0.03992782
#> units9  -0.619279562  0.70919541 -0.18880820 -0.24658536 -1.37847206
#> units10  0.428971412  1.39512141 -0.75484563 -0.68407364 -0.62808398
#> units11 -1.444596011  1.73107841 -0.49139782 -0.10003412 -1.12256280
#> units12 -1.230018808  1.73792333 -0.07047149  0.29416502 -0.53586421
#> units13  0.320979104  0.98660798  1.82338488  2.07578534  2.20647031
#> units14  0.061428532  0.34988556  0.57976556  1.04859655  0.95127341
#> units15  0.369859750 -2.24114145 -0.37272604 -0.66513686 -0.16229587
#> units16  0.412188198 -1.40776694 -0.02559583 -1.15027812 -0.06711963
#> units17 -0.157404387 -1.08354926 -0.38790771 -1.50181082 -0.22246561
#> units18  0.311419152 -1.85787303 -0.35177398 -0.70479026 -0.15349835
#> units19  0.916064656 -1.29505347 -0.07654214 -0.55905812 -0.24965284
#> units20  0.963842737 -0.84154665 -0.59072569 -0.93151452 -0.22736382
#> units21  1.120794747 -0.60268072 -0.45083523 -0.22628807 -0.35846649
#> units22 -0.152003849 -0.08923965  0.07589860 -0.20930266 -0.77536148
#> units23  1.640045179  0.37554892 -0.85708586 -0.37990655 -1.06162251
#> units24 -0.266317210  0.10516333  0.04190799  0.25415481 -2.16522735
#> units25  0.944217058  0.82432163  1.10867350 -0.10748692  1.08533718
#> units26  1.064627655 -0.53010349 -1.42200868 -1.20534184 -0.70419153
#> units27 -0.004229887 -0.52729004 -1.37391907  0.14075976 -1.33895050
#> units28 -0.142825002  0.37659395 -0.48445480  0.03465754 -0.54741639
#>            Zoraspin
#> units1   0.27024496
#> units2   0.60788071
#> units3  -0.17472419
#> units4   1.30350928
#> units5   1.02338930
#> units6   2.13534563
#> units7   0.98751904
#> units8   0.08150135
#> units9  -0.87782667
#> units10 -0.42521674
#> units11 -1.35671295
#> units12 -0.71058207
#> units13  1.55165531
#> units14  1.49370759
#> units15  0.07437018
#> units16  0.20018526
#> units17  0.02166245
#> units18 -1.38528811
#> units19  0.04662180
#> units20 -0.09507075
#> units21 -0.10845618
#> units22 -0.49175387
#> units23 -0.56210304
#> units24 -1.49174730
#> units25  0.89026182
#> units26 -0.75629133
#> units27 -0.56577556
#> units28 -0.71305423