Simulates new data from a given cord object

Simulates new data from a given cord object

Usage

# S3 method for cord
simulate(object, nsim = 1, seed = NULL, newdata = object$obj$data, ...)

Arguments

object

is a cord object, e.g. from output of cord

nsim

Number of simulations, defaults to 1. If nsim > 1, the simulated data will be appended.

seed

Random number seed, defaults to a random seed number.

newdata

A data frame in which to look for X covariates with which to simulate.

...

not used Defaults to the X covariates in the fitted model.

Examples

abund = spider$abund

spider_mod_ssdm = stackedsdm(abund,~1, data = spider$x, ncores=2)
spid_lv_ssdm = cord(spider_mod_ssdm)
simulate(spid_lv_ssdm, nsim=2)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        0        6        0        2        0        0       30        0
#>  [2,]        0        9        0        0        0        1       36        0
#>  [3,]        0        3        0        6        0        0       26        0
#>  [4,]        2        9        5        0        0        0        2        3
#>  [5,]       20        1       30        0        4        1        0       91
#>  [6,]       16        0        0        5        0        0        0       10
#>  [7,]        8       11        0        0        0        1        0       10
#>  [8,]        1       33        0        1        0       17        3        1
#>  [9,]        0       17        0        0        0        1        7        2
#> [10,]        0        7        0        0        1        0        7        0
#> [11,]        1        4        0        2        0        0        3        1
#> [12,]        3        0        0        0        0        0        5        0
#> [13,]        0       12        0        1        0        0        4        6
#> [14,]       58       46        3        0        0        1        0      102
#> [15,]       24       15        0        1        0        1        0       46
#> [16,]        0        0        0        1        0        0       28        1
#> [17,]        9        2       35        0        0        0        0        3
#> [18,]        3        0        0        0        0        0        5        9
#> [19,]        5       13        3        0        0        1        0       28
#> [20,]        9        0        1        0        0        0        0       33
#> [21,]       38        0       19        2        1        4        0      116
#> [22,]        4        3        6        0        0        1        0        6
#> [23,]        0        3        0        0        0        1        2        1
#> [24,]        0        3        0        1        0        0        6        0
#> [25,]        6        4        0        0        0        1        0      123
#> [26,]        5       11        1        0        0       22        0       55
#> [27,]       10        1        0        0        0       15        2      107
#> [28,]        3        0        9        0        0        0        0        3
#> [29,]       14        2        1        0        1        1        1        4
#> [30,]        0       12        0        0        0        3        1        2
#> [31,]       31        0        8        0        0        3        0       27
#> [32,]        0       13        1        0        0        4        3       15
#> [33,]        1       13        0        4        0        4        2        0
#> [34,]       19       12        0        4        0        0        0       32
#> [35,]       19        5        1        1        0        0        0       37
#> [36,]       15        5        1        1        0        6        0      111
#> [37,]       21        7        8        2        0        1        0       72
#> [38,]        0        0        0        0        0        0       11        0
#> [39,]        5        6        1        4        0       10        0       75
#> [40,]        7        5        0        0        4        2        0        8
#> [41,]        0        3        2        0        0        0        1        2
#> [42,]       20        0        6        0        0        0        1       31
#> [43,]        8        0        2        0        0        0        0        0
#> [44,]        2        0        7        0        0        0        0        2
#> [45,]        7        3        2        0        0        3        0        5
#> [46,]        2        1        6        0        0        0        8        1
#> [47,]        0        1        1        0        2        2        0        0
#> [48,]        0        0        0        0        1        0        2        3
#> [49,]        8        0        6        0        0        0        0        0
#> [50,]        1        6        0        0        0        0        3       18
#> [51,]        0        1        0        0        0        1        3        0
#> [52,]       17        0       37        0        0        0        0       59
#> [53,]        2        1        0        0        0        3        0        9
#> [54,]        0        5        0        5        0       31        2        1
#> [55,]        0        2        0        0        0        0        3        0
#> [56,]        0        4        0        3        0        0        1        7
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]       18        7       88       20
#>  [2,]        1        1       33        9
#>  [3,]        0        0       30        1
#>  [4,]        1        1       40        4
#>  [5,]        1        0        1        0
#>  [6,]       12       11       27        3
#>  [7,]        8       14       57        1
#>  [8,]      112       38      105        2
#>  [9,]       19        7       26        5
#> [10,]       27        7       30        4
#> [11,]       10      119       52        4
#> [12,]        0        0        7        0
#> [13,]        6        0       28        9
#> [14,]       20        7       20        3
#> [15,]        4       65        2        0
#> [16,]        0        0       19        0
#> [17,]        1        2        6        0
#> [18,]        0        0       17        3
#> [19,]        2        1       11        1
#> [20,]       22        1       12        0
#> [21,]        2       11        3        0
#> [22,]        1        0        2        0
#> [23,]       10       48       90        2
#> [24,]        0        0       20        1
#> [25,]        1        0        1        0
#> [26,]       73       39       57       13
#> [27,]        9        3       10        2
#> [28,]        0        0        0        0
#> [29,]        1       11       18        3
#> [30,]        4        0       82        8
#> [31,]        0        2        3        0
#> [32,]        4       12       36        9
#> [33,]        6        0       45        3
#> [34,]       13      115       97        5
#> [35,]        0       24       11        0
#> [36,]       21       71       35       18
#> [37,]       62       19       33        3
#> [38,]        0        2       70       14
#> [39,]       50        1       27        4
#> [40,]        0        1        4        0
#> [41,]        0        7       41       18
#> [42,]        0        0        1        0
#> [43,]        0        0        0        0
#> [44,]        0        0        1        1
#> [45,]        7        2       14        1
#> [46,]        0        0        2        1
#> [47,]        2        0       22        7
#> [48,]        3        2       37        1
#> [49,]        0        0        0        0
#> [50,]        0       76       28        6
#> [51,]        0        0       20        2
#> [52,]        0        4        5        0
#> [53,]        6        1       20       12
#> [54,]       83       20       83       30
#> [55,]        1        0       40        4
#> [56,]        0       10        7        2

# \donttest{
# using mvabund
library(mvabund) #for manyglm
abund=mvabund(abund)
spider_mod = manyglm(abund~1)
spid_lv = cord(spider_mod)
simulate(spid_lv)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       12        0        6        0        0        0        1        2
#>  [2,]        7        2        5        0        2        0        1        3
#>  [3,]        1        1        7        0        0        0       19        1
#>  [4,]        0       12        0        0        0        1        4        0
#>  [5,]       45        2        6        0        4        0        0        9
#>  [6,]        0        2        3        0        0        0       15        1
#>  [7,]        0        7        0        0        0        5       10        0
#>  [8,]        0        0        0        5        0        0       13        0
#>  [9,]        0        0        0        0        0        0        1        0
#> [10,]        5        3        0        0        1        7        0       42
#> [11,]        4        1       17        0        0        0        0        1
#> [12,]        5        0        0        0        4        0        2        1
#> [13,]        4        0       28        0        5        0        0       30
#> [14,]        0        0        0        0        0        0       17        0
#> [15,]        6        1        4        0       15        1        1        4
#> [16,]        4       18        2        3        0       12        0       21
#> [17,]       41        0       24        0        0        0        0       24
#> [18,]        0        0        4        0        0        0        4        0
#> [19,]       12        1        1        0        0        0        0        7
#> [20,]       13        0        3        0        0        0        0       42
#> [21,]       11        6        0        0        3        0       10        5
#> [22,]        3        0        2        0       29        0        0       10
#> [23,]        0        0        0        0        0        0        2        3
#> [24,]        4       15        0       13        0        0        0        4
#> [25,]        0        1        0        0        0        0       17        0
#> [26,]       12        0        2        0        1        0        1       11
#> [27,]        0        2        4        0        0        0        6        0
#> [28,]        1        2        0        0        1        1       12        8
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        1       10        1
#>  [2,]        0        0       10        0
#>  [3,]       14        2       45        1
#>  [4,]       26        0       46       18
#>  [5,]        0        2        1        0
#>  [6,]        0        0       12        0
#>  [7,]        0       24       32        9
#>  [8,]        0        0       48        8
#>  [9,]        0        0        0        0
#> [10,]       11        4       43        0
#> [11,]        1        0       25        0
#> [12,]        1        1       31        4
#> [13,]        0        0        0        0
#> [14,]        0        0       30        4
#> [15,]        1        0       14        1
#> [16,]       51       79       66        1
#> [17,]        0        0        0        0
#> [18,]        0        0        1        0
#> [19,]        0        2       18        3
#> [20,]        0        0        0        0
#> [21,]        0        0        1        0
#> [22,]        0        1        4        1
#> [23,]        0        0        3        0
#> [24,]       32        2       51        4
#> [25,]        3        0       92        9
#> [26,]        0        0        0        0
#> [27,]        0        0       22        1
#> [28,]       30       16       42       13

spider_mod_X = manyglm(abund ~ soil.dry + bare.sand, data=spider$x)
spid_lv_X = cord(spider_mod_X)
Xnew = spider$x[1:10,]
simulate(spid_lv_X, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       16        4        1        0        0        1        4       18
#>  [2,]        0        0        1        0        0        0        1        0
#>  [3,]        6       10        0        0        0        9        0       19
#>  [4,]        2        0        0        0        0        0        2        3
#>  [5,]        2        8        0        0        0        1        9       21
#>  [6,]        0        0        0        0        0        0        0        0
#>  [7,]        0        0        1        0        0        3        7        1
#>  [8,]        7        2        0        0        0        1        1       46
#>  [9,]        0        1        0        0        0       14       11        1
#> [10,]        1        0        2        0        0        0        0        0
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        1        0       13        1
#>  [2,]        0        0        2        0
#>  [3,]       73        5       39        1
#>  [4,]        0        0        5        3
#>  [5,]        8       18       28        7
#>  [6,]        0        0       16        5
#>  [7,]        1        0        8        4
#>  [8,]       20       65       41        4
#>  [9,]        3        2       71        4
#> [10,]        0        0        4        6
simulate(spid_lv_X, nsim=2, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        1        0        1        0        0        0        3       13
#>  [2,]        4       61        1        0        0       27        3      130
#>  [3,]        2        6        0        0        0        3        1        1
#>  [4,]        0        2        0        0        0        0        5        0
#>  [5,]        0        9        0        0        0        0        3        1
#>  [6,]        1        6        2        4        0        2        0        5
#>  [7,]        0        8        0        0        0        1        2        0
#>  [8,]        0        1        0        0        0        7       13        0
#>  [9,]        0        2        0        0        0        0       54        0
#> [10,]        1        2        5        0        1        0        0        0
#> [11,]        8        0        1        0        0        0        0       27
#> [12,]        0        0        0        0        0        0        2        0
#> [13,]        6       24        0        2        0        4        0       47
#> [14,]        6       11        1        5        0       11        0       71
#> [15,]        0        5        0        0        0        4        7        0
#> [16,]        0        0        0        0        0        0        2        0
#> [17,]        5       18        0       13        0       40        2        9
#> [18,]        7       19        0        0        0       11        0      130
#> [19,]        4        5        0        0        0        4        0        9
#> [20,]        0        0        0        0        0        0        0        0
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        0        1        0
#>  [2,]       76      229      163       18
#>  [3,]        6      125       61       19
#>  [4,]        7        4       21        4
#>  [5,]        0        3       35        7
#>  [6,]       57        3       86       22
#>  [7,]        0       31       23        1
#>  [8,]       16        8       68        8
#>  [9,]        0        0       37        3
#> [10,]        0        1       11        7
#> [11,]        0        0        3        0
#> [12,]        0        0        1        0
#> [13,]       10       26       51       11
#> [14,]       31       83       88       17
#> [15,]       68       23      171       31
#> [16,]        0        0        1        2
#> [17,]        9       83      128       24
#> [18,]       10        4       23        5
#> [19,]        2       18       38        0
#> [20,]        0        0        1        1

spider_mod_X_ssdm = stackedsdm(abund, formula_X = ~. -bare.sand, data = spider$x, ncores=2)
spid_lv_X_ssdm = cord(spider_mod_X_ssdm)
simulate(spid_lv_X_ssdm, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        8        3        1        0        0        5        1       13
#>  [2,]        3        0        1        0        0       19        4       14
#>  [3,]       17        5        0        0        0        5        0       30
#>  [4,]       13        2        0        0        0       11        0       64
#>  [5,]        5       15        0        1        0        9        1       23
#>  [6,]        0       10        0        1        0        2        2        3
#>  [7,]        1        1        1        3        0        8        0       23
#>  [8,]        0        0        0        0        0        3       23        0
#>  [9,]        4        2        1        0        0        0        0       89
#> [10,]        6        0        1        0        0        0        0       18
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        1       23        5
#>  [2,]        3       20      106       14
#>  [3,]        6       68       33        2
#>  [4,]       11       28       39        2
#>  [5,]      199       25      201       44
#>  [6,]        7       17       48        8
#>  [7,]       64       51       75        7
#>  [8,]        0        0       21        2
#>  [9,]        5        3       13        1
#> [10,]        0        0        3        0
# }