Simulates new data from a given cord object

Simulates new data from a given cord object

Usage

# S3 method for cord
simulate(object, nsim = 1, seed = NULL, newdata = object$obj$data, ...)

Arguments

object

is a cord object, e.g. from output of cord

nsim

Number of simulations, defaults to 1. If nsim > 1, the simulated data will be appended.

seed

Random number seed, defaults to a random seed number.

newdata

A data frame in which to look for X covariates with which to simulate.

...

not used Defaults to the X covariates in the fitted model.

Examples

abund = spider$abund

spider_mod_ssdm = stackedsdm(abund,~1, data = spider$x, ncores=2)
spid_lv_ssdm = cord(spider_mod_ssdm)
simulate(spid_lv_ssdm, nsim=2)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        0        6        3        0        0        0        4       17
#>  [2,]        0        7        0        0        0        0        6        0
#>  [3,]        0        2        0        0        0        0        2        4
#>  [4,]        1        0        0        5        2        0        2        0
#>  [5,]        8        2        0        0        0        1        0       24
#>  [6,]        0        4        0        1        0       13       18        0
#>  [7,]        3        3        0        0        0        0        0       10
#>  [8,]        2        0        1        0        1        0        1        0
#>  [9,]        1        0       29        0        0        0        0        2
#> [10,]        3        0        0        0        1        0        4        0
#> [11,]        0        7        0        0        0        2        0       14
#> [12,]        0        1        0        0        1        7        0        7
#> [13,]        0        7        0        0        0        1        4        0
#> [14,]        0        0        1        0        6        0        5        3
#> [15,]        0       40        0        2        0       29       19        0
#> [16,]        3        2       13        0        3        0        0       12
#> [17,]        1        1        2        0        2        0        1        2
#> [18,]       14       12        4        0        0        0        7       25
#> [19,]        0        3        0        0        0        0        3        0
#> [20,]       59        2       17        0        3        1        0        4
#> [21,]        1        0        0        0        0        0        5        1
#> [22,]        0        0        0        0        0        0        2        0
#> [23,]        5       16        0        4        5        0       11       30
#> [24,]        2        6        0        0        0        0        0       25
#> [25,]        5        2        0        0        0       19        0       45
#> [26,]        2       14        4        1        0       91        0       53
#> [27,]        4        0        0        0        0        0        0        0
#> [28,]        0        4        0        0        0        1        1        3
#> [29,]        6       14        1        2        0        3        0       34
#> [30,]        0        0        0        0        0        4        9        4
#> [31,]        0       25        0       13        0        9        0       11
#> [32,]        0        0        0        0        1        0        0       27
#> [33,]        0        3        0        2        0        1        1        0
#> [34,]        4        0        0        0        0        6        1        2
#> [35,]        3        0        1        0        0        0       13       10
#> [36,]        2        3        0        0       24        1        0        0
#> [37,]        0       15        0        0        0        2       12        3
#> [38,]        0       67        0        4        0       25        0       62
#> [39,]       11        0        2        0        0        0        2        2
#> [40,]        8        0       45        0       32        0        0        0
#> [41,]        5        0        8        0        1        0        4        3
#> [42,]        0        1        2        0        0        0       10        0
#> [43,]        9       18        0        2        0       49        0      154
#> [44,]        0        3        0        1        0        4        5        6
#> [45,]        0        0        0        0        0        0        2        0
#> [46,]        4        1        0        0        0        0        7        0
#> [47,]        0        0        5        0        1        0        3        0
#> [48,]       33        0       33        0        4        0        0       29
#> [49,]        1       22        4       16        2        0        1        1
#> [50,]       10        5        5        1       10        0        0       19
#> [51,]        0       25        0        0        0        5       18        0
#> [52,]       17       11        2        5        0       45        0       69
#> [53,]        0        0        0        0        0        0       13        3
#> [54,]        4        1        2        0       18        2        1       15
#> [55,]        5        3        0        3        0        0        2        3
#> [56,]       17        2       36        0        6        4        0       17
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        3       12       68       10
#>  [2,]        5        0       45        2
#>  [3,]        2        0       14        3
#>  [4,]        0        0       23        7
#>  [5,]       48        1       23       10
#>  [6,]       32        5      161       17
#>  [7,]        2       54       25        3
#>  [8,]        0        0        2        0
#>  [9,]        0        0        1        0
#> [10,]        2        0        5        1
#> [11,]       11       55       66       25
#> [12,]        2        8       28        1
#> [13,]        0        1       45        6
#> [14,]        0        0       11        0
#> [15,]        7        7      264       64
#> [16,]        0        0        3        2
#> [17,]        0        6       21        3
#> [18,]        0        0       21        1
#> [19,]        0        0       48       27
#> [20,]        0        0        1        0
#> [21,]        0        0        3        3
#> [22,]        0        0        1        1
#> [23,]      118       53       50        1
#> [24,]        5       23       25        5
#> [25,]       30        6       21        1
#> [26,]       78       59      100       16
#> [27,]        0        0        1        0
#> [28,]        2        2       49       18
#> [29,]       55       51       38        6
#> [30,]        1       15       47        5
#> [31,]       33       42      138       18
#> [32,]        0        6        6        1
#> [33,]        0        0       39        1
#> [34,]        2        1       15        1
#> [35,]        0        1       13        1
#> [36,]        0        0       13        6
#> [37,]       18        1       45       24
#> [38,]      105       78      153       65
#> [39,]        0        0        0        0
#> [40,]        0        0        0        0
#> [41,]        0        0        2        1
#> [42,]        0       12       46        2
#> [43,]       38      524      114       22
#> [44,]       15       25      121       19
#> [45,]        0        0        3        0
#> [46,]        1        0       18        0
#> [47,]        0        0        0        0
#> [48,]        0        7        0        0
#> [49,]      106       43       62        5
#> [50,]        2        6       12        0
#> [51,]       51       21       86       25
#> [52,]      142      207      109       31
#> [53,]        0        0       22       18
#> [54,]       25        0       14        1
#> [55,]        6        8       23        1
#> [56,]        0        0        0        0

# \donttest{
# using mvabund
library(mvabund) #for manyglm
abund=mvabund(abund)
spider_mod = manyglm(abund~1)
spid_lv = cord(spider_mod)
simulate(spid_lv)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        0        9        0        6        0       25       14        1
#>  [2,]        0        0        0        0        3        0       29        0
#>  [3,]        0       39        0        1        0       40       23        9
#>  [4,]        9       35        0        6        0        3        1      136
#>  [5,]        7        0        0        0        7        4        0        7
#>  [6,]        0        2        0        4        0        3        8       32
#>  [7,]        0        6        0        0        0        2        0        1
#>  [8,]        0        1        0        0        0        0        6        1
#>  [9,]        0        0        0        1        0        2        2        0
#> [10,]        0        1        2        0        0        7        1        0
#> [11,]       11        8       28        0        0        3        0        9
#> [12,]        0        5        1        0        0        2        2       47
#> [13,]        1        2        0        0        0        1       13        4
#> [14,]        0       15        0        0        0        1       11        4
#> [15,]        0        0        0        0        0        0        4        3
#> [16,]        6        0        1        0        0        0        0        4
#> [17,]        0        0        4        0        0        0        0        0
#> [18,]        2        0        0        0        0        0        2        0
#> [19,]        8        3        8        0        2        1        0      134
#> [20,]        7       15        0        3        0        1        0       26
#> [21,]       25        0        6        0        0        0        0        9
#> [22,]        5        0       36        0        0        0        0        0
#> [23,]        3        0        0        1        0        0        5       21
#> [24,]       15        9        1        1        0       14        0       23
#> [25,]        5        8        2        1        0        0        0        4
#> [26,]        0        9        4        2        0        0        2        0
#> [27,]        1       12        0        2        0        0        0        3
#> [28,]        0       21        0        1        0        2        6        1
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]       50       89      136       73
#>  [2,]        0        0        2        0
#>  [3,]      105        8      133       23
#>  [4,]        2      208       74        2
#>  [5,]       13       12       32        3
#>  [6,]       18      275      150       10
#>  [7,]        8        1      128        6
#>  [8,]        0        0       50        5
#>  [9,]        2        0       35        1
#> [10,]        1        8       40        4
#> [11,]        1        0       10        4
#> [12,]       13        1       38        6
#> [13,]        0        0       26        1
#> [14,]        8       46       40        0
#> [15,]        0        0        0        0
#> [16,]        3        0        2        0
#> [17,]        2        1       17        3
#> [18,]       16        0       21        7
#> [19,]        0        2        3        0
#> [20,]       11        2       39        3
#> [21,]        0        0        0        0
#> [22,]        0        0        0        0
#> [23,]        0        0       34        4
#> [24,]       17        1       45        7
#> [25,]        0        0       13       16
#> [26,]        5       45       43        7
#> [27,]        1        5       43        6
#> [28,]      116       11       72        8

spider_mod_X = manyglm(abund ~ soil.dry + bare.sand, data=spider$x)
spid_lv_X = cord(spider_mod_X)
Xnew = spider$x[1:10,]
simulate(spid_lv_X, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        4        3        1        0        0        0        0        2
#>  [2,]        5       14        0        0        0        1        4       74
#>  [3,]        1        5        0        0        0        0        0        7
#>  [4,]       13        5        2        0        0        4        3       21
#>  [5,]        0        1        0        4        0        0       15        1
#>  [6,]        0        3        1        7        0       37        0        1
#>  [7,]        0        1        0        0        0        1        1        1
#>  [8,]        6       22        0        1        0       17        0       24
#>  [9,]        1        0        0        0        0        0        2        0
#> [10,]        0        0        8        0        1        0        0        6
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        0        7        0
#>  [2,]        2       73       16        1
#>  [3,]        0        8       14        1
#>  [4,]       11        7       17        0
#>  [5,]        1        1       22        3
#>  [6,]      358        3       46        8
#>  [7,]        0        0       64        5
#>  [8,]        2       24       41        9
#>  [9,]        0        0        1        0
#> [10,]        0        0        5        7
simulate(spid_lv_X, nsim=2, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        7        7        0        0        0        5        7       19
#>  [2,]        1        3        0        1        0        8        1       24
#>  [3,]        5        6        0        0        0        0        0       13
#>  [4,]        4        1        0        0        0        0        7        0
#>  [5,]        0        0        0        0        0        9        4        0
#>  [6,]        1        2        0        0        0       15        0        8
#>  [7,]        1        7        0        6        0       18       44       33
#>  [8,]        0        0        0        0        0        0       19        0
#>  [9,]       18       14        0        0        0        9        1       63
#> [10,]        1        0        1        0        1        0        0        0
#> [11,]        5        2        2        0        0       11        3        4
#> [12,]        0        6        0        2        0        1        0       21
#> [13,]        5       46        0        1        0       36       22        3
#> [14,]        7       11        0        0        0       10        0       36
#> [15,]        1        2        0        0        0        0       10        6
#> [16,]        0        0        1        4        0        3        1        4
#> [17,]        2       22        0        1        0        6        6       24
#> [18,]        8       10        0        0        0       10       11        5
#> [19,]       10        5        1        0        0        2        0       74
#> [20,]        0        0        4        0        0        1        1        0
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]       15       75       69       13
#>  [2,]        8        9       24        4
#>  [3,]        1        1       18        0
#>  [4,]        0        0        3        3
#>  [5,]        3       11       60        9
#>  [6,]       33       45       60       29
#>  [7,]       46      250      107       16
#>  [8,]        0        2        6        2
#>  [9,]       29       45       81        8
#> [10,]        3        0       13        7
#> [11,]       15        3       54        8
#> [12,]        5       99       36        6
#> [13,]       52      138      122       20
#> [14,]       20       46       52        1
#> [15,]        0        0        1        0
#> [16,]      100        9      141       46
#> [17,]        0       75       54        8
#> [18,]       54      174       86       13
#> [19,]        0       25       26        0
#> [20,]        0        0        5        3

spider_mod_X_ssdm = stackedsdm(abund, formula_X = ~. -bare.sand, data = spider$x, ncores=2)
spid_lv_X_ssdm = cord(spider_mod_X_ssdm)
simulate(spid_lv_X_ssdm, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       19       11        0        0        1        3        0       81
#>  [2,]        1        3        0        4        0       67        9       13
#>  [3,]       15       36        2        0        0       18        1       62
#>  [4,]       10        0        0        0        0       15        0       97
#>  [5,]        0       20        0        7        0        7        2        1
#>  [6,]        0       13        0        1        0        3        0        7
#>  [7,]        0       21        0        5        0       14        1       10
#>  [8,]        0        0        0        0        0        3       41        1
#>  [9,]        2        0        1        0        0        0        0       48
#> [10,]        5        1        4        0        0        0        0       10
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        3       27       16        0
#>  [2,]       69       72       82       59
#>  [3,]       25       55       33        6
#>  [4,]        0       24       15        1
#>  [5,]       26       39       98       11
#>  [6,]        7        4       45        7
#>  [7,]      137       57      105       54
#>  [8,]        0        0       12        2
#>  [9,]        0        0        0        0
#> [10,]        1        1        9        0
# }