Simulates new data from a given cord object

Simulates new data from a given cord object

Usage

# S3 method for cord
simulate(object, nsim = 1, seed = NULL, newdata = object$obj$data, ...)

Arguments

object

is a cord object, e.g. from output of cord

nsim

Number of simulations, defaults to 1. If nsim > 1, the simulated data will be appended.

seed

Random number seed, defaults to a random seed number.

newdata

A data frame in which to look for X covariates with which to simulate.

...

not used Defaults to the X covariates in the fitted model.

Examples

abund = spider$abund

spider_mod_ssdm = stackedsdm(abund,~1, data = spider$x, ncores=2)
spid_lv_ssdm = cord(spider_mod_ssdm)
simulate(spid_lv_ssdm, nsim=2)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        5        0       16        0       33        0        0        7
#>  [2,]       48        0       10        0        8        0        0       31
#>  [3,]        1        2        0        0        0        0        1        0
#>  [4,]        0        0        0        2        0        0        4        2
#>  [5,]        0        0        0        0        0        0       19        0
#>  [6,]        0        8        1        0        0        2        1        0
#>  [7,]       33        0        9        0       23        0        0       10
#>  [8,]        0       12        0        0        0        0        7        1
#>  [9,]       30        5       22        0        0        3        1       20
#> [10,]        4        0        1        0        0        0        0        9
#> [11,]        0        0        0        0        0        0        0        0
#> [12,]       17        4        0        0       30        0        1       10
#> [13,]        0       38        0       13        0       11       21        3
#> [14,]        1        6        0        0        8        1        0       33
#> [15,]        0        3        0        0        0        0       26        0
#> [16,]        7        1        0        0        0        0        1        3
#> [17,]        7        1        2        0       22        0        2        1
#> [18,]        0        2        1        0        0        0        7        1
#> [19,]        7        3        1        0        0        4        0      104
#> [20,]        1        3        0        0        1        3        0        9
#> [21,]        8        2        0        1        0        1        4        6
#> [22,]        0       20        0        0        0        2        2        0
#> [23,]        2        0       13        0        0        0        1        6
#> [24,]        0        8        0        5        0        8        0       13
#> [25,]        3        0        1        0        0        0        0        0
#> [26,]        1        3        0        0        6        0        0       19
#> [27,]        0        1        0        0        0        0        5        1
#> [28,]       19        0        2        0        0        0        2       10
#> [29,]        1        0        0        1        1        0        1        1
#> [30,]        3       14        0        2        0        5        7        1
#> [31,]        4        0       32        0        0        0        0       12
#> [32,]        1       28        0        0        0        1        3        2
#> [33,]        4        1       37        1        0        0       11        0
#> [34,]       19        1        0        0        2        3        0       23
#> [35,]        1        0       15        0        5        0        0        0
#> [36,]        3        7        0        0        0        7        0       11
#> [37,]        1        0        8        0        1        0        1        5
#> [38,]        6        0       19        0        1        0        0        1
#> [39,]        6        2        5        0       20        0        1        2
#> [40,]        0        3        0        0        0       34        0       12
#> [41,]       21       12        0       13        0        1        0       46
#> [42,]       58        1       22        0        6        0        0       12
#> [43,]        5       12        0        0        0        0        6        1
#> [44,]        4       10        3       15        0       12        3       21
#> [45,]        0        0        0        0        0        0        2        0
#> [46,]        0       11        0        0        0        9        1        0
#> [47,]        1       12        0        0        0       35       13        1
#> [48,]        7        3        0        1        0        6        0        9
#> [49,]        0        3        0        0        0        1        9        1
#> [50,]        8        0        1        0        0        0        0       17
#> [51,]        1        3        7        1       27        0        0       22
#> [52,]        0        0        0        0        0        0        3        3
#> [53,]        0        0        2        0        0        0        5        5
#> [54,]       20        0        3        1        0        0        1       26
#> [55,]        0        2       17        0        0        1        1        0
#> [56,]        7        5        0        7        0        0        0       36
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        0        0        0
#>  [2,]        0        6        1        0
#>  [3,]        0        3       13        1
#>  [4,]        0        0        4        1
#>  [5,]        0        0       28        3
#>  [6,]        0        3      105        8
#>  [7,]        0        0        0        0
#>  [8,]        1        3       40       13
#>  [9,]        0        0        3        0
#> [10,]        0        0        1        0
#> [11,]        0        0        7        2
#> [12,]        1        0        5        0
#> [13,]       35      220      226       34
#> [14,]        1        3       12        0
#> [15,]        2        0       70        3
#> [16,]       12        0       11        2
#> [17,]        1        0        8        0
#> [18,]        1        0       29        0
#> [19,]        2        9       10        6
#> [20,]        1        0        8       17
#> [21,]        3       87       56        3
#> [22,]        9      133      148       23
#> [23,]        0        0        0        0
#> [24,]       88      104       98       12
#> [25,]        0        0        0        0
#> [26,]        8       16       29        3
#> [27,]        5        0       28        1
#> [28,]        0        0        2        0
#> [29,]        0        0        8        0
#> [30,]       60        7      124       11
#> [31,]        0        0        5        1
#> [32,]       70       25      136       20
#> [33,]        0        0        5        0
#> [34,]        0        2        3        0
#> [35,]        0        0        1        0
#> [36,]        8        4       20        2
#> [37,]        0        0        0        0
#> [38,]        0        0        0        0
#> [39,]        0        0        2        0
#> [40,]       72       96      168       33
#> [41,]       11       11       27        5
#> [42,]        0        4        1        0
#> [43,]       22       34      141       30
#> [44,]       16       26       69       10
#> [45,]        0        0        8        0
#> [46,]        2       30       83       26
#> [47,]        8        0       37        6
#> [48,]        0        0        1        1
#> [49,]       13        0       40       11
#> [50,]        0        2        3        0
#> [51,]        3        0        5        2
#> [52,]        0        8       30        3
#> [53,]        0        0        3        1
#> [54,]        0        0        3        0
#> [55,]        0        0        3        0
#> [56,]        5        7       21        1

# \donttest{
# using mvabund
library(mvabund) #for manyglm
abund=mvabund(abund)
spider_mod = manyglm(abund~1)
spid_lv = cord(spider_mod)
simulate(spid_lv)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        0        7        0        0        0        0        4        0
#>  [2,]       12        0       21        0       11        0        2       73
#>  [3,]        0       10        0        1        0        0       44        0
#>  [4,]       28        2        5        0        0        9        0       11
#>  [5,]        0        5        0        1        0        0        8       21
#>  [6,]       14        3        3        7        0        0        5       40
#>  [7,]        0       22        0        0        0       10        1        1
#>  [8,]        7        5        0        0        2        8        1       22
#>  [9,]        2        3        0        9        0        0        0        3
#> [10,]       18        0       27        0        7        0        0       15
#> [11,]        2        1        0        0        4        0       17        0
#> [12,]        4        0        4        0        9        0        6        6
#> [13,]        0        1        0        0        0        0       12        2
#> [14,]       25        8        0        0        0        6        0       34
#> [15,]        1        6        6        1        0        1       13        0
#> [16,]        1        0        0        0        0        6        0        1
#> [17,]        0       10        0        3        0        0        7        0
#> [18,]        1        8        0        0        0        0        0        5
#> [19,]        7        0        8        0        0        0        4       19
#> [20,]        0        4        0        0        0        0        9        1
#> [21,]        0        0        7        0       10        0        5        0
#> [22,]        0        0        0        0        0        1       38        1
#> [23,]        3       86        0       10        0       84        0       36
#> [24,]        1        0        3        0        4        0        0        2
#> [25,]        3        0        4        0        0        0        0       43
#> [26,]        0        0        0        0        0        0       12        0
#> [27,]       34        2        1        0        0        0        3       57
#> [28,]        9        6        0        0        0       13        0       73
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        6       41        2
#>  [2,]        1        1        1        0
#>  [3,]       87       44      166       34
#>  [4,]        0       32       20        0
#>  [5,]        1        2       14        0
#>  [6,]        0        0        5        0
#>  [7,]        4       42      114       17
#>  [8,]       23        7       36        4
#>  [9,]       25        4       57        8
#> [10,]        0        0        5        0
#> [11,]        0        0       10        2
#> [12,]        0        0        0        0
#> [13,]        3        0       19        4
#> [14,]        7       76       47        2
#> [15,]       34        2       53       13
#> [16,]        0        0       14        2
#> [17,]       14       49      134       28
#> [18,]        6        0       58        4
#> [19,]        0        0        7        1
#> [20,]        1       11       67        3
#> [21,]        0        0        0        0
#> [22,]        2        0       19        0
#> [23,]      264      349      270       27
#> [24,]        0        0        1        0
#> [25,]        0        0        1        0
#> [26,]        0        0       24        0
#> [27,]        1        2       14        0
#> [28,]       13       30       56        3

spider_mod_X = manyglm(abund ~ soil.dry + bare.sand, data=spider$x)
spid_lv_X = cord(spider_mod_X)
Xnew = spider$x[1:10,]
simulate(spid_lv_X, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        0        5        0        0        0        6        0        2
#>  [2,]        5        3        0        2        0        2        0       96
#>  [3,]       18       43        2        1        0       13        7      146
#>  [4,]        2        2        0        0        0        2       10        3
#>  [5,]        0        0        0        0        0        0       19        0
#>  [6,]        2        5        0        3        0        5        5        9
#>  [7,]        1       10        1        1        0        2        1       13
#>  [8,]        2        1        0        0        0        0        4        0
#>  [9,]        1        0        0        0        0        0        0        0
#> [10,]        2        1        7        0        0        0        0        6
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        1        1       57        3
#>  [2,]        7       12       28        9
#>  [3,]       69      160       61       12
#>  [4,]        2        0       27        4
#>  [5,]        1        0       35        6
#>  [6,]        2        2       52        3
#>  [7,]        2       13      118       17
#>  [8,]       15        1       77        4
#>  [9,]        0        1        6        0
#> [10,]        0        0        4        3
simulate(spid_lv_X, nsim=2, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        8        1        1        0        0        1        2        0
#>  [2,]        2       29        0        6        0        0       12        8
#>  [3,]        2        6        0        0        0        0        6        1
#>  [4,]        1        0        0        0        0        2       15        0
#>  [5,]        0        1        0        0        0        2        3        0
#>  [6,]        3       26        0        0        0       45        0       21
#>  [7,]        2       30        0        0        0       10        6       10
#>  [8,]        0        3        1        0        0        2       26        1
#>  [9,]        3        0        0        0        0        0        1        0
#> [10,]        1        2        8        0        1        0        0        3
#> [11,]        4       18        1        0        0       30        2       45
#> [12,]        2       12        1        0        0        0        1       12
#> [13,]        0        0        0        0        0        0       24        0
#> [14,]        1        0        0        0        0        0        2        5
#> [15,]        1       25        0        3        0        5        0       26
#> [16,]        0        0        1        1        0        2        0        2
#> [17,]        1        6        0        2        0        0        5        0
#> [18,]        5        5        0        0        0       11        1       31
#> [19,]        2        2        2        1        0       11        0       39
#> [20,]        0        0        3        0        0        4        0        5
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        3       17        1
#>  [2,]        1       69       44        1
#>  [3,]        0        2        5        3
#>  [4,]        3        0       25       10
#>  [5,]        0       14       38       13
#>  [6,]        2       24       74       16
#>  [7,]        0        0       98       14
#>  [8,]        0        0       40        5
#>  [9,]        0        0        1        0
#> [10,]        0        1       16        5
#> [11,]       44      104       63       12
#> [12,]        0        0       38        0
#> [13,]        0        0        6        0
#> [14,]        0        1       25        5
#> [15,]       32        3       25        5
#> [16,]        0        0       15       25
#> [17,]       39      446      115       26
#> [18,]        3       13       23        1
#> [19,]       13       19       33       11
#> [20,]        2        0       10        7

spider_mod_X_ssdm = stackedsdm(abund, formula_X = ~. -bare.sand, data = spider$x, ncores=2)
spid_lv_X_ssdm = cord(spider_mod_X_ssdm)
simulate(spid_lv_X_ssdm, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       14        2        4        0        0        5        0       50
#>  [2,]        2       24        0        0        0       35        0       18
#>  [3,]       15        1        2        0        0       22        0        2
#>  [4,]        6        7        1        3        0       28        0       31
#>  [5,]        3        4        0        0        0        9        0       20
#>  [6,]        0       17        0        7        0        1        0        4
#>  [7,]        2       11        1        0        0        6        1        8
#>  [8,]        0        0        0        0        0        3       37        0
#>  [9,]        9        0        1        1        0        0        0       70
#> [10,]        5        5        0        0        0        0        0        7
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        5       44       19        1
#>  [2,]       54       81      117       51
#>  [3,]        3       24       17        1
#>  [4,]       39       83       81        5
#>  [5,]       19       38      103        9
#>  [6,]       32        7      116       62
#>  [7,]       50      124       53       24
#>  [8,]        1        1       14        3
#>  [9,]        0        0        3        0
#> [10,]        2        1        8        0
# }