Simulates new data from a given cord object

Simulates new data from a given cord object

Usage

# S3 method for cord
simulate(object, nsim = 1, seed = NULL, newdata = object$obj$data, ...)

Arguments

object

is a cord object, e.g. from output of cord

nsim

Number of simulations, defaults to 1. If nsim > 1, the simulated data will be appended.

seed

Random number seed, defaults to a random seed number.

newdata

A data frame in which to look for X covariates with which to simulate.

...

not used Defaults to the X covariates in the fitted model.

Examples

abund = spider$abund

spider_mod_ssdm = stackedsdm(abund,~1, data = spider$x, ncores=2)
spid_lv_ssdm = cord(spider_mod_ssdm)
simulate(spid_lv_ssdm, nsim=2)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        0        1        0        0        0        0        6        1
#>  [2,]        3       13        3        5        0       19        0       47
#>  [3,]        0       49        0        3        0       11        7        0
#>  [4,]        0       33        0        1        0        5        3        0
#>  [5,]        0       16        2        1        0        1        1        6
#>  [6,]        3       10        1        0        3        0        2        1
#>  [7,]        0        5        0        0        0        7        0        0
#>  [8,]        0        9        0        2        0       13       19        0
#>  [9,]       12        0        0        2        0        0        0       12
#> [10,]        7        4        0        0        0       10        0        0
#> [11,]        1        3        0        0        0        0        2        0
#> [12,]        9        6        1        0        0        1        0       15
#> [13,]        0        5        0        1        4        0       13        1
#> [14,]       15       20       21        1        0       41        0      290
#> [15,]        0        0        0        2        0        0        7        0
#> [16,]        3        4        0        0        0        4        0       12
#> [17,]        0        0        0        0        0        0       12        0
#> [18,]       29        4        0        0        0        0        2       44
#> [19,]        0        0        0        0        0        0        2        1
#> [20,]        0        1        0        1        0        4        4        0
#> [21,]        0        1        0        0        0        1        4        0
#> [22,]        4        1        1        1        0        0        0        5
#> [23,]        0        4        0        1        0        3        4        2
#> [24,]        5        5        0        0        0        3        1        1
#> [25,]        0        0        4        0        0        0        0        0
#> [26,]        0        4        0        0        5        1        3        1
#> [27,]        0       13        0        1        0        0       13        8
#> [28,]        2       35        0        6        0       20       16       31
#> [29,]       12       13        7        0        2        0        3       27
#> [30,]        8        0        0        0        6        0        4        1
#> [31,]        2        0        0        0        0        0        1        0
#> [32,]        4        9        0        0        0       50        0       24
#> [33,]        5        7        1        0        0       28        0       31
#> [34,]       41        7        7        1        0        0        0       97
#> [35,]        0        0        1        0        0        9        2        0
#> [36,]       20       18       26        0        0       28        0       49
#> [37,]        4        2        2        0        3        0        0       33
#> [38,]        1        0        0        0        0        0        2        0
#> [39,]        0        3        0        0        0        0        7        0
#> [40,]        1        0        8        0        0        0        2        1
#> [41,]        2        0        0        0        0        5        4        9
#> [42,]        0        4        0        2        1        1        6        0
#> [43,]        4        0        2        0        0        1        0        4
#> [44,]        7        6        3        0        0        6        0        4
#> [45,]        6        3        0        0        1        0        0       77
#> [46,]       23        2        2        1        0       13        0       27
#> [47,]        0        0        0        0        0       15        5       14
#> [48,]       19       10        9        0        0        4        3       10
#> [49,]        0        7        0        0        0        4       12        0
#> [50,]        0        1        0        1        0        8       27        0
#> [51,]        0       34        0        9        0       30        4        4
#> [52,]        6        8        3        0        0        3        0       18
#> [53,]        0        5        0        0        0        4        7        3
#> [54,]        3        0        0        1        0        0        1       16
#> [55,]        1        1        0        3        0        0        0        2
#> [56,]        0        0        2        0       31        0        2       14
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        0       30        2
#>  [2,]       11       21       40        3
#>  [3,]      119      187      239       71
#>  [4,]        5       52       98       30
#>  [5,]        2       28       31        0
#>  [6,]        0        0        4        0
#>  [7,]        5        2       37        2
#>  [8,]        2       11      138       12
#>  [9,]        0        0        2        1
#> [10,]        5        6       36        2
#> [11,]        0        0        9        2
#> [12,]        2       17       12        5
#> [13,]        9        1       65        7
#> [14,]       41      140       83        6
#> [15,]        1        0       50        2
#> [16,]       42        3       51        2
#> [17,]        2        0       31        2
#> [18,]        3        3       19        0
#> [19,]        0        1       16       13
#> [20,]        0       19       71       19
#> [21,]        1        0       33        2
#> [22,]        0        0        5        0
#> [23,]        0        8       65        3
#> [24,]        3        3       15        0
#> [25,]        0        0        8        0
#> [26,]       58       95      142       19
#> [27,]       12       94       82       11
#> [28,]      118      138      209       30
#> [29,]        0        1        4        0
#> [30,]        0        1        1        0
#> [31,]        0        0       19        0
#> [32,]       19       89       54        3
#> [33,]       65       74       44        2
#> [34,]        6       54        8        7
#> [35,]        0        1       40        0
#> [36,]       21      108       84        6
#> [37,]        9        9       18        1
#> [38,]        0        0        7        0
#> [39,]        0        0       65        2
#> [40,]        0        0        1        0
#> [41,]        5       36       19        1
#> [42,]       10        1       76       22
#> [43,]        0        0        5        0
#> [44,]        2       15       41        3
#> [45,]        6        9        8        0
#> [46,]       95        0       10        0
#> [47,]       14        4       43       11
#> [48,]        0        1       12        7
#> [49,]       23       63      141       55
#> [50,]        7       17       90        6
#> [51,]      123      107      146       50
#> [52,]        0        0        5       24
#> [53,]       74       15      115       52
#> [54,]        3        0       15        4
#> [55,]       15       10       32        9
#> [56,]        0        2       21       13

# \donttest{
# using mvabund
library(mvabund) #for manyglm
abund=mvabund(abund)
spider_mod = manyglm(abund~1)
spid_lv = cord(spider_mod)
simulate(spid_lv)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        8        6        0        0        0        1        0       12
#>  [2,]        1        0        0        0        0        0        0        6
#>  [3,]        9        0        0        0        0        0        1        9
#>  [4,]       21        1        1        0        1        0        0       20
#>  [5,]        0        0        0        0        0        0       18        0
#>  [6,]       23        1       18        4        0        0        0        5
#>  [7,]        2        0        2        0        0        0        0        0
#>  [8,]        0        4        0        0        0        0        0        1
#>  [9,]        0       16        0        1        0        1        6       12
#> [10,]        0       13        0        3        0        5        3        2
#> [11,]        0        2        0        0        0        2       11        6
#> [12,]        0        0        0        0        0        0        3        0
#> [13,]        2       13        0        0        0        3       10        9
#> [14,]        1        0        1        0        0        0        0        2
#> [15,]        4        0        0        0        0        0        2       70
#> [16,]       16        5        0        1        0        0        0       19
#> [17,]        0        0        0        0        4        0        3        1
#> [18,]        0        0        0        0        0        0        1        0
#> [19,]        1        0        2        0        0        0        1        0
#> [20,]        0        4        0        0        0        2        2        0
#> [21,]        4        6        0        0        0        0        0        9
#> [22,]        0       11        3        3        0       23        0       25
#> [23,]        0        1        0        0        0        0       21        0
#> [24,]       21        6        1        0        0        0        0       67
#> [25,]       15        2        2        0        2        2        0       17
#> [26,]        4        2        2        0        0        0       12        5
#> [27,]        1       10        0        0        0        0        9        1
#> [28,]        3        2        0        0        0        6        1        6
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0       11       18        1
#>  [2,]        0        8       20        3
#>  [3,]        0        0        0        1
#>  [4,]        0        1        0        0
#>  [5,]        0        0        1        1
#>  [6,]        0        0        1        0
#>  [7,]        0        0        5        2
#>  [8,]        0        7       25       43
#>  [9,]       70       41       65        7
#> [10,]      291       96      214       38
#> [11,]        3        9      127        5
#> [12,]        0        0       25        1
#> [13,]       17        1       65        2
#> [14,]        0        0        3        4
#> [15,]       29        4       17        0
#> [16,]        0        1       20        0
#> [17,]        0        0        4        0
#> [18,]        0        0        2        0
#> [19,]        0        0        1        0
#> [20,]        4        4      155        8
#> [21,]        3        0        4        0
#> [22,]       21        4       68       15
#> [23,]        4        0       63       14
#> [24,]        5        0        4        2
#> [25,]        0        0        2        0
#> [26,]       23        7       29        5
#> [27,]        4        0       47        9
#> [28,]        7       22       25        3

spider_mod_X = manyglm(abund ~ soil.dry + bare.sand, data=spider$x)
spid_lv_X = cord(spider_mod_X)
Xnew = spider$x[1:10,]
simulate(spid_lv_X, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        4        5        1        0        0        7        0       13
#>  [2,]        0        0        0        0        0        0       17        2
#>  [3,]        5        6        0        0        0        3       11        3
#>  [4,]        4       11        0        0        0        0        8       12
#>  [5,]        0        4        1        0        0       12        8       24
#>  [6,]        0       16        1        1        0       53        0       18
#>  [7,]        0        6        0        5        0        0        0       21
#>  [8,]        0        4        1        0        0        4       16        1
#>  [9,]        7        5        0        0        0        0        6       59
#> [10,]        1        0        2        0        0        0        0        0
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        5        9       58        5
#>  [2,]        0        0       13        0
#>  [3,]        0       10       18        0
#>  [4,]        2        0       19        7
#>  [5,]       59      196       76       16
#>  [6,]      222       55      135      125
#>  [7,]        0        0       38        0
#>  [8,]        0        1       16        3
#>  [9,]        0        0        6        0
#> [10,]       10        0       20       29
simulate(spid_lv_X, nsim=2, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        1        0        0        0        0        0        0        0
#>  [2,]        3       25        0        3        0       31        0        2
#>  [3,]        0        0        0        0        0       15        5        0
#>  [4,]        0        0        0        0        0        0       37        0
#>  [5,]        4        9        0        2        0        1        0        9
#>  [6,]        0        1        0        0        0        0        0       10
#>  [7,]        0        1        0        0        0       16        2        1
#>  [8,]        1        2        0        0        0       14       17        3
#>  [9,]       15        7        1        0        0        4        0       30
#> [10,]        0        0        4        0        0        0        0        0
#> [11,]       17        5        3        1        0        0        8       64
#> [12,]        0        0        0        0        0        0        1        1
#> [13,]       10        1        3        1        0       23       11       56
#> [14,]        6        5        0        0        0        0        0        0
#> [15,]        4        6        0        1        0        0        0       78
#> [16,]        0        0        0        0        0        0        0        3
#> [17,]        0        9        0        0        0        0       16        0
#> [18,]        8        8        0        0        0        8        5      138
#> [19,]        1        0        0        0        0        0        1        0
#> [20,]        1        2        8        0        1        1        0        7
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        0        2        2
#>  [2,]       84       36       71        7
#>  [3,]        1        0       24        2
#>  [4,]        0        0        0        0
#>  [5,]        2       77       50        2
#>  [6,]       25        0       39       15
#>  [7,]        0        3       42        1
#>  [8,]        5        0       20        5
#>  [9,]       18       29       65        8
#> [10,]        0        0        4        0
#> [11,]        1       22        4        0
#> [12,]        0        0        1        0
#> [13,]        7       10       64        2
#> [14,]        2        0       19        4
#> [15,]       14      104       25        3
#> [16,]       14        0        6        1
#> [17,]        0        0       52       14
#> [18,]       11       33       47        1
#> [19,]        0        0        2        0
#> [20,]        2        0       10        3

spider_mod_X_ssdm = stackedsdm(abund, formula_X = ~. -bare.sand, data = spider$x, ncores=2)
spid_lv_X_ssdm = cord(spider_mod_X_ssdm)
simulate(spid_lv_X_ssdm, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       25        0        1        0        1       10        0      103
#>  [2,]        0        3        0        1        0       42        4        6
#>  [3,]        7        0        1        0        0        3        0        6
#>  [4,]       18        0        1        0        0       31        0      164
#>  [5,]        4        4        1        1        0        2        0        5
#>  [6,]        0       11        0        7        0        0        2        2
#>  [7,]        1        9        0        9        0       12        1        5
#>  [8,]        0        2        0        0        0        4       44        1
#>  [9,]        7        0        0        0        0        0        0      113
#> [10,]       13        6        4        0        0        0        0       20
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        1       11       10        0
#>  [2,]       27      107       48        3
#>  [3,]        3       12       28        3
#>  [4,]       10       37       65       16
#>  [5,]        7       17       20        1
#>  [6,]        9       16       50        9
#>  [7,]       96      237       68        9
#>  [8,]        1        2       70        6
#>  [9,]        0        1        5        0
#> [10,]        2        0       11        2
# }