Simulates new data from a given cord object

Simulates new data from a given cord object

Usage

# S3 method for cord
simulate(object, nsim = 1, seed = NULL, newdata = object$obj$data, ...)

Arguments

object

is a cord object, e.g. from output of cord

nsim

Number of simulations, defaults to 1. If nsim > 1, the simulated data will be appended.

seed

Random number seed, defaults to a random seed number.

newdata

A data frame in which to look for X covariates with which to simulate.

...

not used Defaults to the X covariates in the fitted model.

Examples

abund = spider$abund

spider_mod_ssdm = stackedsdm(abund,~1, data = spider$x, ncores=2)
spid_lv_ssdm = cord(spider_mod_ssdm)
simulate(spid_lv_ssdm, nsim=2)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       14        0        0        0        4        0        1        1
#>  [2,]        0        2        0        0        0        3       67        0
#>  [3,]       22       12        0        0        2        0        3       15
#>  [4,]        0        0        0        0        0        0       13       11
#>  [5,]        0        0        0        0        0        0       25        0
#>  [6,]       22        3        4        4        0        9        0      117
#>  [7,]        1        0        0        0        0        0        0       11
#>  [8,]       11        2        0        2        0        1        0       24
#>  [9,]        0        2        0        0        3       13        0        0
#> [10,]        0        0        0        0        0        0       54        0
#> [11,]       20        0       23        0        2        0        0        0
#> [12,]        0        0        1        0        0        0        7        0
#> [13,]        0        0        0        0        0        1        3        5
#> [14,]        2       32        0        0        0        2        2       77
#> [15,]        0        0        0        0        0        0       12        1
#> [16,]        0        0        0        0        0        1       20        0
#> [17,]       35        0        1        1        0        1        0       31
#> [18,]       17        4        5        1        8        0        0       24
#> [19,]        0       17        0        0        0        4        1        7
#> [20,]        0        0        0        0        1        0        3        0
#> [21,]        1        2       10        0        0        0        0       18
#> [22,]        0        3        0        3        0        3        3        0
#> [23,]       12        0        1        0       17        0        4        0
#> [24,]       18        0       14        0        7        0        1       46
#> [25,]        2        1       16        0        0        0        0        4
#> [26,]        7        0        6        0        0        0        0       35
#> [27,]        0        3        0        1        0        0        4        8
#> [28,]        0        0        0        8        0        1        5        0
#> [29,]       12        0        1        0        1        0        0        2
#> [30,]        0        4        0        0        0        6        5       32
#> [31,]       25        1        4        0        0        0        1       16
#> [32,]        0        0        2        0        0        0        7        0
#> [33,]        3        0        0        0        0        0        2        7
#> [34,]       25        6        6        1        0        2        0       91
#> [35,]        0        7        0        2        0        0       42        0
#> [36,]        1        0        1        0        0        0        2        0
#> [37,]       13        0        1        0        0        6        0        2
#> [38,]        8        1        0        0        0        0        0        9
#> [39,]        3        1        0        1        1        0        8        3
#> [40,]        2        0        7        0        0        0        3        0
#> [41,]        0       37        0        1        0       48        1       10
#> [42,]        1        9        0        0        1        6        1        2
#> [43,]        0        6        0        0        0        4        0        4
#> [44,]        0       19        1        2        0        9        1       45
#> [45,]        0        4        0        5        0       44        5       21
#> [46,]        3        1        0        0        0        0        2       11
#> [47,]        0        4        0        0        0        0        5        4
#> [48,]        3        2        1        0        0        0        1        6
#> [49,]        0       40        0        1        0       24        3        2
#> [50,]        1       19        0        0        0        0        0        6
#> [51,]       10       19        0        0        0       18        4       35
#> [52,]        6        0       59        0       29        0        0        5
#> [53,]        0        2        2        0        0        0       12        2
#> [54,]        6        0        0        0        0        2        1        3
#> [55,]       11        5        4        0        0        0        0       21
#> [56,]        9        5        2        0        0        0        0       30
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        0        0        0
#>  [2,]        8       21       82       20
#>  [3,]        9       58       17        3
#>  [4,]        0        0       48        2
#>  [5,]        0        0       12        3
#>  [6,]        3       84       47        5
#>  [7,]        3        1       18        4
#>  [8,]        3       13        9        0
#>  [9,]        6        0       20       21
#> [10,]        0        0       29        0
#> [11,]        0        0        0        0
#> [12,]        0        0       36        0
#> [13,]       10       18       59        5
#> [14,]        6       12       46        7
#> [15,]        0        0        2        8
#> [16,]        1        0       45        1
#> [17,]        8        0        2        3
#> [18,]        3        0       10        0
#> [19,]        1       34      104        1
#> [20,]        0        0       22        2
#> [21,]        0        0        2        0
#> [22,]        2        1       60       12
#> [23,]        0        0        1        4
#> [24,]        1        0        3        0
#> [25,]        0        0        3        0
#> [26,]        1        1       13        0
#> [27,]       21       73       55       13
#> [28,]        5        0       48        1
#> [29,]        0        0        0        0
#> [30,]       86       16      134       37
#> [31,]        0        0        4        1
#> [32,]        0        0        7        0
#> [33,]       10       56       13        4
#> [34,]        8       23       23        2
#> [35,]        5        0       53       39
#> [36,]        0        0        0        0
#> [37,]        0        0        9        4
#> [38,]        3        0        4        0
#> [39,]        1        0        5        9
#> [40,]        0        0        0        0
#> [41,]      186      148      126       37
#> [42,]       24        5       40        6
#> [43,]        4       38       23        4
#> [44,]        7        0       97       12
#> [45,]       81       88      104       45
#> [46,]        0        0        8        3
#> [47,]        2        0       14        1
#> [48,]        0        0       14        0
#> [49,]       90      195      158       15
#> [50,]       43       48       83       24
#> [51,]       47      152       66       11
#> [52,]        0        0        0        0
#> [53,]        0        0       48        2
#> [54,]        2        0       18        0
#> [55,]        0        9        4        1
#> [56,]       11        5       31        4

# \donttest{
# using mvabund
library(mvabund) #for manyglm
abund=mvabund(abund)
spider_mod = manyglm(abund~1)
spid_lv = cord(spider_mod)
simulate(spid_lv)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        1       13        0        6        0        3        0       13
#>  [2,]       13        0       17        0       24        0        1        9
#>  [3,]        7        0        6        0       14        0        1        5
#>  [4,]        0       39        0        1        0        9       13       26
#>  [5,]        0        1        0        0        1        1       20        7
#>  [6,]        2        0        3        0        1        0        3        1
#>  [7,]       15        1        1        0        0        0        1       42
#>  [8,]        0       12        0        0        0        6        2        5
#>  [9,]        3        3        0        1        1       12        0        7
#> [10,]        0        0        0        0        0        0       18        2
#> [11,]        2        0       10        0        0        0        1        4
#> [12,]        1        0        3        0        0        0        2        1
#> [13,]        3        5        0        0        0        1        2        0
#> [14,]        0        7        0        1        0        1        9        8
#> [15,]        2        1        3        0        0        2        0       35
#> [16,]        0        1        0        1        0        0       14        7
#> [17,]        4        0        6        0        2        1        0        9
#> [18,]        0        0        2        0        0        0        3        1
#> [19,]        0        1        0        0        5        0        0       27
#> [20,]       14        6        0        0        0       11        0        6
#> [21,]        0        8        1        0        0        1       21        6
#> [22,]        1        1        0        0        0        0        0        0
#> [23,]        3        0        4        0       14        0        0       12
#> [24,]       11        0        7        0       18        1        0      112
#> [25,]        1        0        0        1       13        1       21        0
#> [26,]       10        1        3        0        0        0        1       33
#> [27,]        7       14        1        0        0       25        9       33
#> [28,]        4        7        8        1        0       21        0       16
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]       12        4       39       10
#>  [2,]        0        0        0        0
#>  [3,]        0        0        0        0
#>  [4,]       31       28      199       18
#>  [5,]        0        0       34       17
#>  [6,]        1        0       23        1
#>  [7,]        2       54       40        2
#>  [8,]        9        8       49        7
#>  [9,]        7        0       81        6
#> [10,]        0        0        4        0
#> [11,]        7       54       35        3
#> [12,]        3        0        6        0
#> [13,]        7        8       45        1
#> [14,]        1        0       34        6
#> [15,]        3       30       40        6
#> [16,]        0        0       28        3
#> [17,]        0        0        6        0
#> [18,]        1        1       20        2
#> [19,]        0        2       33       12
#> [20,]        1        2       19        0
#> [21,]       17       16       14        0
#> [22,]        0        9       31        1
#> [23,]        0        0        2        0
#> [24,]        6        0        1        0
#> [25,]        2        1       30       23
#> [26,]        0        0        6        0
#> [27,]        0        1        7        0
#> [28,]       16       15       17        0

spider_mod_X = manyglm(abund ~ soil.dry + bare.sand, data=spider$x)
spid_lv_X = cord(spider_mod_X)
Xnew = spider$x[1:10,]
simulate(spid_lv_X, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       11        2        0        0        0        1        2        3
#>  [2,]        3        1        0        4        0        0       13       44
#>  [3,]       13        4        1        3        0        5        2       63
#>  [4,]        2        1        1        0        0        4       24       11
#>  [5,]        7       67        0        6        0       68        0       46
#>  [6,]        0        7        1        0        0        0        0        3
#>  [7,]        0        4        0        0        0        5       12        0
#>  [8,]        5       12        0        2        0        0       22        7
#>  [9,]        7        1        0        0        0        0        0       14
#> [10,]        9        2        4        1        0        3        0       30
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        3        0       21        2
#>  [2,]       12       47       78        3
#>  [3,]       12       77       54        4
#>  [4,]        1        8       26       10
#>  [5,]       40      148      139       21
#>  [6,]       67        0       96       20
#>  [7,]       33      186       68       14
#>  [8,]        1       12       16        0
#>  [9,]        0        4        2        0
#> [10,]       12        1       28       12
simulate(spid_lv_X, nsim=2, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]        4        4        0        1        0       30       14       13
#>  [2,]        6        2        0        0        0       10        4       25
#>  [3,]        5       12        0        0        0        0        0       16
#>  [4,]        1        0        0        0        0        0       10        0
#>  [5,]        2        1        0        0        0        0        1       27
#>  [6,]        1       11        0        0        0       17        0        1
#>  [7,]        0        1        0        3        0        0       10        0
#>  [8,]        6        7        0        0        0       11       28       34
#>  [9,]        1        1        0        0        0        0       15        1
#> [10,]        0        0        1        0        3        0        0        2
#> [11,]        8        0        2        0        0        0        5       47
#> [12,]        0        0        0        0        0        0        3        0
#> [13,]        0        1        0        0        0        0       30        2
#> [14,]        6        2        0        0        0        0        0      124
#> [15,]        0        3        0        2        0        1        1        0
#> [16,]        0        2        0        3        0        1        0        3
#> [17,]        7       19        0        1        0        4        3       42
#> [18,]        2       15        0        1        0       10        2       17
#> [19,]        8        0        0        3        0        6       28        1
#> [20,]        0        3        1        0        0        1        0        0
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]       39       75       95       18
#>  [2,]        2       72       30        1
#>  [3,]        0       14       38        0
#>  [4,]        0        0       11        0
#>  [5,]        1       41       41        0
#>  [6,]      217       21      130       75
#>  [7,]        2        0       29        5
#>  [8,]        4       39       61        7
#>  [9,]        0        2       30        1
#> [10,]        0        0        4        1
#> [11,]        0        1       12        1
#> [12,]        0        0        5        3
#> [13,]        0        0       29        4
#> [14,]        0        4        2        0
#> [15,]        0        1       50       10
#> [16,]       48        1       63        6
#> [17,]      113      172       53       22
#> [18,]       38       27       92        3
#> [19,]       14       13       32        7
#> [20,]       14        0       21       13

spider_mod_X_ssdm = stackedsdm(abund, formula_X = ~. -bare.sand, data = spider$x, ncores=2)
spid_lv_X_ssdm = cord(spider_mod_X_ssdm)
simulate(spid_lv_X_ssdm, newdata = Xnew)
#>       Alopacce Alopcune Alopfabr Arctlute Arctperi Auloalbi Pardlugu Pardmont
#>  [1,]       11        0        2        0        1        5        0        9
#>  [2,]        1       14        0        0        0       63        5       20
#>  [3,]        7        4        1        0        0       18        0       33
#>  [4,]       13        8        1        2        0        6        0       36
#>  [5,]        0        0        0        3        0        9        0        4
#>  [6,]        0        4        0        0        0        2        1        3
#>  [7,]        2        8        0        5        0       14        1       37
#>  [8,]        1        6        0        0        0        2       12        1
#>  [9,]        5       11        0        1        0        0        0       40
#> [10,]        3        4        0        0        0        0        0       11
#>       Pardnigr Pardpull Trocterr Zoraspin
#>  [1,]        0        9       10        3
#>  [2,]       27       68      146       12
#>  [3,]        2       42       30        5
#>  [4,]       11       40       15        3
#>  [5,]      191       20       61        8
#>  [6,]        1        4       36       14
#>  [7,]      186      121      169       78
#>  [8,]        1        0       16        2
#>  [9,]        7        2       16        2
#> [10,]        0        0        4        1
# }